Agrupar por población especial y casos

Función que agrupa los datos de la enfermedad o evento por población especial y casos

Uso

agrupar_pob_especial(data_event, nomb_col = "poblacion", porcentaje = TRUE)

Argumentos

data_event

Un `data.frame` que contiene los datos de la enfermedad o evento

nomb_col

Un `character` (cadena de caracteres) con el nombre de la columna de los datos de la enfermedad o evento que contiene las poblaciones especiales; su valor por defecto es `"poblacion"`

porcentaje

Un `boolean` (TRUE o FALSE) que indica si es necesario agregar un porcentaje de casos como una columna; su valor por defecto es `TRUE`

Valor

Un `data.frame` con los datos de la enfermedad o evento agrupados por poblaciones especiales y casos

Ejemplos

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_pob_especial(data_event = data_limpia,
                     nomb_col = "poblacion",
                     porcentaje = TRUE)
#>     poblacion  casos
#> 1  gp_discapa 127389
#> 2  gp_desplaz 127238
#> 3  gp_migrant 127035
#> 4  gp_carcela 127412
#> 5   gp_gestan 127173
#> 6  gp_indigen 127461
#> 7  gp_pobicfb 127416
#> 8  gp_mad_com 127476
#> 9  gp_desmovi 127477
#> 10 gp_psiquia 127471
#> 11 gp_vic_vio 127425
#> 12   gp_otros  64687