Agrupar por sexo, semana epidemiológica y casos

Función que agrupa los datos de enfermedades por sexo, semana epidemiológica y número de casos

Uso

agrupar_sex_semanaepi(
  data_event,
  cols_sex = c("sexo", "semana"),
  porcentaje = TRUE
)

Argumentos

data_event

Un `data.frame` que contiene los datos de la enfermedad o evento

cols_sex

Un `character` (cadena de caracteres) o `array (arreglo) de character` con el nombre de la(s) columna(s) que contienen el sexo y las semanas epidemiológicas; su valor por defecto es `c("sexo", "semana")`

porcentaje

Un `boolean` (TRUE o FALSE) que indica si se debe agregar una columna con el porcentaje de casos; su valor por defecto es `TRUE`

Valor

Un `data.frame` con los datos de la enfermedad o evento agrupados por sexo, semana epidemiológica y número de casos

Ejemplos

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_sex_semanaepi(data_event = data_limpia,
                      cols_sex = c("sexo", "semana"),
                      porcentaje = TRUE)
#> # A tibble: 15 × 7
#>    sexo  semana cod_eve nombre_evento ano   casos porcentaje
#>    <chr> <chr>  <chr>   <chr>         <chr> <int>      <dbl>
#>  1 F     01     210     DENGUE        2020      1       2.22
#>  2 F     03     210     DENGUE        2020     14      31.1 
#>  3 F     04     210     DENGUE        2020      1       2.22
#>  4 F     05     210     DENGUE        2020      1       2.22
#>  5 F     08     210     DENGUE        2020      1       2.22
#>  6 F     11     210     DENGUE        2020      1       2.22
#>  7 F     14     210     DENGUE        2020      1       2.22
#>  8 F     15     210     DENGUE        2020      1       2.22
#>  9 M     02     210     DENGUE        2020      1       2.22
#> 10 M     03     210     DENGUE        2020     18      40   
#> 11 M     09     210     DENGUE        2020      1       2.22
#> 12 M     10     210     DENGUE        2020      1       2.22
#> 13 M     12     210     DENGUE        2020      1       2.22
#> 14 M     13     210     DENGUE        2020      1       2.22
#> 15 M     16     210     DENGUE        2020      1       2.22