Agrupar por la clasificacion inicial del caso

Función que agrupa los casos por la clasificacion inicial del caso

agrupar_tipo_caso(data_event, nomb_cols = "tip_cas")

Argumentos

data_event

Un `data.frame` que contiene los datos de la enfermedad o evento

nomb_cols

Un `character` (cadena de caracteres) con el nombre de las columna(s) en los datos de la enfermedad o evento que contiene la clasificacion inicial del caso; su valor por defecto es `"tip_cas"`

Valor

Un `data.frame` con los datos de la enfermedad o evento agrupados por la clasificacion inicial del caso y/u otras variables como años

Examples

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_tipo_caso(data_event = data_limpia,
                  nomb_cols = "tip_cas")
#> # A tibble: 3 × 5
#>   tip_cas cod_eve nombre_evento casos nombre_tip_cas
#>   <chr>   <chr>   <chr>         <int> <named list>  
#> 1 2       210     DENGUE        47838 <chr [1]>     
#> 2 3       210     DENGUE        15591 <chr [1]>     
#> 3 5       210     DENGUE          313 <chr [1]>