Clean department codes of the disease data

Function that cleans the department codes of the disease data

Usage

clean_depto_codes(disease_data, col_data_codes, geo_data, col_geo_codes)

Arguments

disease_data

The disease data

col_data_codes

Column name in the disease data that contains the department codes

geo_data

Geographical data that includes the departments codes

col_geo_codes

Column name in the geographical data than contains the department codes

Value

Clean department codes of the disease data

Examples

geo_codes <- import_geo_codes()
disease_data <- import_linelist_disease_year(2019, "DENGUE")
clean_depto_codes(disease_data, 
                  col_data_codes = "cod_dpto_o", 
                  geo_codes, 
                  col_geo_codes = "c_digo_departamento")
#>         CONSECUTIVE COD_EVE    FEC_NOT SEMANA  ANO    COD_PRE COD_SUB EDAD
#>      1:     6389811     210 2019-05-19     20 2019 5200100283       1    4
#>      2:     6389656     210 2019-09-17     37 2019 5000600169       1   31
#>      3:     6389809     210 2019-08-30     35 2019 5242700696       1    5
#>      4:     6389810     210 2019-08-30     35 2019 5242700696       1    5
#>      5:     6389812     210 2019-09-02     35 2019 5242700696       1    6
#>     ---                                                                   
#> 123637:     7096713     210 2019-12-23     51 2019 1800107466       1    2
#> 123638:     6584651     210 2019-11-01     43 2019 1100185079      80   21
#> 123639:     6584634     210 2019-12-23     51 2019 7000100365       1    2
#> 123640:     6584642     210 2019-10-03     39 2019 7331900990       1   21
#> 123641:     6584648     210 2019-07-05     27 2019 7326800794       1    1
#>         UNI_MED nacionalidad nombre_nacionalidad SEXO COD_PAIS_O COD_DPTO_O
#>      1:       1          170            COLOMBIA    F        170         52
#>      2:       1          170            COLOMBIA    F        170         50
#>      3:       1          170            COLOMBIA    F        170         52
#>      4:       1          170            COLOMBIA    M        170         52
#>      5:       1          170            COLOMBIA    F        170         52
#>     ---                                                                    
#> 123637:       1          170            COLOMBIA    F        170         18
#> 123638:       1          170            COLOMBIA    F        170         25
#> 123639:       1          170            COLOMBIA    M        170         70
#> 123640:       1          170            COLOMBIA    F        170         73
#> 123641:       1          170            COLOMBIA    M        170         73
#>         COD_MUN_O AREA LOCALIDAD       CEN_POBLA                       VEREDA
#>      1:       427    2  SIN DATO DOCE DE OCTUBRE                     SIN DATO
#>      2:         6    1   ACACIAS        SIN DATO                     SIN DATO
#>      3:       427    1     MAGUI        SIN DATO                     SIN DATO
#>      4:       427    1     MAGUI        SIN DATO                     SIN DATO
#>      5:       427    3  SIN DATO        SIN DATO                       PLAYON
#>     ---                                                                      
#> 123637:         1    1 FLORENCIA        SIN DATO                     SIN DATO
#> 123638:       488    1  SIN DATO        SIN DATO                     SIN DATO
#> 123639:         1    1  SIN DATO        SIN DATO                     SIN DATO
#> 123640:       319    1     GUAMO        SIN DATO                     SIN DATO
#> 123641:       268    3  SIN DATO        SIN DATO AGUA BLANCA SEC. DE LA DULCE
#>                        BAR_VER OCUPACION TIP_SS COD_ASE PER_ETN GRU_POB
#>      1:               SIN DATO      9997      S  ESS062       5      NA
#>      2: PORTALES DE SAN CARLOS      9996      C  EPS013       6      NA
#>      3:                PLAYITA      9997      S  ESS118       5      NA
#>      4:             EL PARAISO      9999      S  ESS062       5      NA
#>      5:               SIN DATO      9997      S  ESS062       5      NA
#>     ---                                                                
#> 123637:                 CENTRO      9999      S  EPSS44       6      NA
#> 123638:               SIN DATO      9997      P  RES003       6      NA
#> 123639:        SIN INFORMACION      9999      C  EPS002       6      NA
#> 123640:             LIBERTADOR      9997      C  EPS044       6      NA
#> 123641:               SIN DATO      9999      C  EPS044       6      NA
#>         nom_grupo estrato GP_DISCAPA GP_DESPLAZ GP_MIGRANT GP_CARCELA GP_GESTAN
#>      1:                 1          2          2          2          2         2
#>      2:                 3          2          2          2          2         2
#>      3:                 1          2          2          2          2         2
#>      4:                 1          2          2          2          2         2
#>      5:                 1          2          2          2          2         2
#>     ---                                                                        
#> 123637:                NA          2          2          2          2         2
#> 123638:                 3          2          2          2          2         2
#> 123639:                NA          2          2          2          2         2
#> 123640:                 2          2          2          2          2         2
#> 123641:                NA          2          2          2          2         2
#>         sem_ges GP_INDIGEN GP_POBICFB GP_MAD_COM GP_DESMOVI GP_PSIQUIA
#>      1:      NA          2          2          2          2          2
#>      2:      NA          2          2          2          2          2
#>      3:      NA          2          2          2          2          2
#>      4:      NA          2          2          2          2          2
#>      5:      NA          2          2          2          2          2
#>     ---                                                               
#> 123637:      NA          2          2          2          2          2
#> 123638:      NA          2          2          2          2          2
#> 123639:      NA          2          2          2          2          2
#> 123640:      NA          2          2          2          2          2
#> 123641:      NA          2          2          2          2          2
#>         GP_VIC_VIO GP_OTROS fuente COD_DPTO_R COD_MUN_R COD_DPTO_N COD_MUN_N
#>      1:          2        1      1         52       427         52     52001
#>      2:          2        1      1         50         6         50     50006
#>      3:          2        1      1         52       427         52     52427
#>      4:          2        1      1         52       427         52     52427
#>      5:          2        1      1         52       427         52     52427
#>     ---                                                                     
#> 123637:          2        1      1         18         1         18     18001
#> 123638:          2        1      1         11         1         11     11001
#> 123639:          2        1      1         70         1         70     70001
#> 123640:          2        1      1         73       319         73     73319
#> 123641:          2        1      1         73       268         73     73268
#>            FEC_CON    INI_SIN TIP_CAS PAC_HOS    FEC_HOS CON_FIN FEC_DEF AJUSTE
#>      1: 2019-05-18 2019-05-13       2       1 2019-05-18       1    <NA>      3
#>      2: 2019-09-17 2019-09-14       2       2       <NA>       1    <NA>      7
#>      3: 2019-08-30 2019-08-28       2       2       <NA>       1    <NA>      0
#>      4: 2019-08-30 2019-08-27       2       2       <NA>       1    <NA>      0
#>      5: 2019-09-02 2019-08-31       2       2       <NA>       1    <NA>      0
#>     ---                                                                        
#> 123637: 2019-12-21 2019-12-17       2       1 2019-12-21       1    <NA>      3
#> 123638: 2019-10-26 2019-10-24       3       1 2019-10-26       1    <NA>      0
#> 123639: 2019-12-23 2019-12-19       2       2       <NA>       1    <NA>      3
#> 123640: 2019-10-03 2019-09-28       2       1 2019-10-03       1    <NA>      0
#> 123641: 2019-07-05 2019-07-02       2       2       <NA>       1    <NA>      7
#>          FECHA_NTO CER_DEF CBMTE    FEC_AJU FM_FUERZA  FM_UNIDAD FM_GRADO
#>      1: 2014-09-21      NA       2019-08-02        NA       <NA>         
#>      2: 1987-11-27      NA       2020-06-17        NA       <NA>         
#>      3: 2013-09-24      NA       2019-09-02        NA       <NA>         
#>      4: 2014-06-27      NA       2019-09-02        NA       <NA>         
#>      5: 2012-10-13      NA       2019-09-04        NA       <NA>         
#>     ---                                                                  
#> 123637: 2016-12-28      NA       2020-01-02        NA       <NA>         
#> 123638: 1998-06-27      NA       2019-11-05         3 1100185079      066
#> 123639: 2017-01-29      NA       2020-01-08        NA       <NA>         
#> 123640: 1998-07-20      NA       2019-10-07        NA       <NA>         
#> 123641: 2017-09-01      NA       2020-06-17        NA       <NA>         
#>         confirmados Estado_final_de_caso             nom_est_f_caso
#>      1:           1                    3 Confirmado por laboratorio
#>      2:           0                    2                   Probable
#>      3:           0                    2                   Probable
#>      4:           0                    2                   Probable
#>      5:           0                    2                   Probable
#>     ---                                                            
#> 123637:           1                    3 Confirmado por laboratorio
#> 123638:           1                    3 Confirmado por laboratorio
#> 123639:           1                    3 Confirmado por laboratorio
#> 123640:           0                    2                   Probable
#> 123641:           0                    2                   Probable
#>         Departamento_ocurrencia Municipio_ocurrencia Departamento_residencia
#>      1:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      2:                    META              ACACIAS                    META
#>      3:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      4:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      5:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>     ---                                                                     
#> 123637:                 CAQUETA            FLORENCIA                 CAQUETA
#> 123638:            CUNDINAMARCA                 NILO                  BOGOTA
#> 123639:                   SUCRE            SINCELEJO                   SUCRE
#> 123640:                  TOLIMA                GUAMO                  TOLIMA
#> 123641:                  TOLIMA              ESPINAL                  TOLIMA
#>         Municipio_residencia Nombre_evento Departamento_Notificacion
#>      1:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      2:              ACACIAS        DENGUE                      META
#>      3:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      4:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      5:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>     ---                                                             
#> 123637:            FLORENCIA        DENGUE                   CAQUETA
#> 123638:               BOGOTA        DENGUE                    BOGOTA
#> 123639:            SINCELEJO        DENGUE                     SUCRE
#> 123640:                GUAMO        DENGUE                    TOLIMA
#> 123641:              ESPINAL        DENGUE                    TOLIMA
#>         Municipio_notificacion
#>      1:                  PASTO
#>      2:                ACACIAS
#>      3:          MAGUI (PAYAN)
#>      4:          MAGUI (PAYAN)
#>      5:          MAGUI (PAYAN)
#>     ---                       
#> 123637:              FLORENCIA
#> 123638:                 BOGOTA
#> 123639:              SINCELEJO
#> 123640:                  GUAMO
#> 123641:                ESPINAL