Estandarizar códigos geográficos de los datos de una enfermedad o evento

Función que estandariza los códigos geográficos de los datos de una enfermedad o evento

Usage

estandarizar_geo_cods(data_event)

Arguments

data_event

Un data frame que contiene los datos de una enfermedad o evento

Value

Un data frame que contiene los códigos geográficos estandarizados de los datos de una enfermedad o evento

Examples

data_event <- import_data_event(2019, "DENGUE")
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
data_event <- limpiar_encabezado(data_event)
estandarizar_geo_cods(data_event = data_event)
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
#>         consecutive cod_eve    fec_not semana  ano       cod_pre cod_sub edad
#>      1:     6389811     210 2019-05-19     20 2019 2.569191e-314       1    4
#>      2:     6389656     210 2019-09-17     37 2019 2.470625e-314       1   31
#>      3:     6389809     210 2019-08-30     35 2019 2.590238e-314       1    5
#>      4:     6389810     210 2019-08-30     35 2019 2.590238e-314       1    5
#>      5:     6389812     210 2019-09-02     35 2019 2.590238e-314       1    6
#>     ---                                                                      
#> 123637:     7096713     210 2019-12-23     51 2019 8.893713e-315       1    2
#> 123638:     6584651     210 2019-11-01     43 2019 5.435637e-315      80   21
#> 123639:     6584634     210 2019-12-23     51 2019 3.458509e-314       1    2
#> 123640:     6584642     210 2019-10-03     39 2019 3.622440e-314       1   21
#> 123641:     6584648     210 2019-07-05     27 2019 3.619921e-314       1    1
#>         uni_med nacionalidad nombre_nacionalidad sexo cod_pais_o cod_dpto_o
#>      1:       1          170            COLOMBIA    F        170         52
#>      2:       1          170            COLOMBIA    F        170         50
#>      3:       1          170            COLOMBIA    F        170         52
#>      4:       1          170            COLOMBIA    M        170         52
#>      5:       1          170            COLOMBIA    F        170         52
#>     ---                                                                    
#> 123637:       1          170            COLOMBIA    F        170         18
#> 123638:       1          170            COLOMBIA    F        170         25
#> 123639:       1          170            COLOMBIA    M        170         70
#> 123640:       1          170            COLOMBIA    F        170         73
#> 123641:       1          170            COLOMBIA    M        170         73
#>         cod_mun_o area localidad       cen_pobla                       vereda
#>      1:       427    2  SIN DATO DOCE DE OCTUBRE                     SIN DATO
#>      2:       006    1   ACACIAS        SIN DATO                     SIN DATO
#>      3:       427    1     MAGUI        SIN DATO                     SIN DATO
#>      4:       427    1     MAGUI        SIN DATO                     SIN DATO
#>      5:       427    3  SIN DATO        SIN DATO                       PLAYON
#>     ---                                                                      
#> 123637:       001    1 FLORENCIA        SIN DATO                     SIN DATO
#> 123638:       488    1  SIN DATO        SIN DATO                     SIN DATO
#> 123639:       001    1  SIN DATO        SIN DATO                     SIN DATO
#> 123640:       319    1     GUAMO        SIN DATO                     SIN DATO
#> 123641:       268    3  SIN DATO        SIN DATO AGUA BLANCA SEC. DE LA DULCE
#>                        bar_ver ocupacion tip_ss cod_ase per_etn gru_pob
#>      1:               SIN DATO      9997      S  ESS062       5      NA
#>      2: PORTALES DE SAN CARLOS      9996      C  EPS013       6      NA
#>      3:                PLAYITA      9997      S  ESS118       5      NA
#>      4:             EL PARAISO      9999      S  ESS062       5      NA
#>      5:               SIN DATO      9997      S  ESS062       5      NA
#>     ---                                                                
#> 123637:                 CENTRO      9999      S  EPSS44       6      NA
#> 123638:               SIN DATO      9997      P  RES003       6      NA
#> 123639:        SIN INFORMACION      9999      C  EPS002       6      NA
#> 123640:             LIBERTADOR      9997      C  EPS044       6      NA
#> 123641:               SIN DATO      9999      C  EPS044       6      NA
#>         nom_grupo estrato gp_discapa gp_desplaz gp_migrant gp_carcela gp_gestan
#>      1:                 1          2          2          2          2         2
#>      2:                 3          2          2          2          2         2
#>      3:                 1          2          2          2          2         2
#>      4:                 1          2          2          2          2         2
#>      5:                 1          2          2          2          2         2
#>     ---                                                                        
#> 123637:                NA          2          2          2          2         2
#> 123638:                 3          2          2          2          2         2
#> 123639:                NA          2          2          2          2         2
#> 123640:                 2          2          2          2          2         2
#> 123641:                NA          2          2          2          2         2
#>         sem_ges gp_indigen gp_pobicfb gp_mad_com gp_desmovi gp_psiquia
#>      1:      NA          2          2          2          2          2
#>      2:      NA          2          2          2          2          2
#>      3:      NA          2          2          2          2          2
#>      4:      NA          2          2          2          2          2
#>      5:      NA          2          2          2          2          2
#>     ---                                                               
#> 123637:      NA          2          2          2          2          2
#> 123638:      NA          2          2          2          2          2
#> 123639:      NA          2          2          2          2          2
#> 123640:      NA          2          2          2          2          2
#> 123641:      NA          2          2          2          2          2
#>         gp_vic_vio gp_otros fuente cod_dpto_r cod_mun_r cod_dpto_n cod_mun_n
#>      1:          2        1      1         52       427         52     52001
#>      2:          2        1      1         50       006         50     50006
#>      3:          2        1      1         52       427         52     52427
#>      4:          2        1      1         52       427         52     52427
#>      5:          2        1      1         52       427         52     52427
#>     ---                                                                     
#> 123637:          2        1      1         18       001         18     18001
#> 123638:          2        1      1         11       001         11     11001
#> 123639:          2        1      1         70       001         70     70001
#> 123640:          2        1      1         73       319         73     73319
#> 123641:          2        1      1         73       268         73     73268
#>            fec_con    ini_sin tip_cas pac_hos    fec_hos con_fin fec_def ajuste
#>      1: 2019-05-18 2019-05-13       2       1 2019-05-18       1    <NA>      3
#>      2: 2019-09-17 2019-09-14       2       2       <NA>       1    <NA>      7
#>      3: 2019-08-30 2019-08-28       2       2       <NA>       1    <NA>      0
#>      4: 2019-08-30 2019-08-27       2       2       <NA>       1    <NA>      0
#>      5: 2019-09-02 2019-08-31       2       2       <NA>       1    <NA>      0
#>     ---                                                                        
#> 123637: 2019-12-21 2019-12-17       2       1 2019-12-21       1    <NA>      3
#> 123638: 2019-10-26 2019-10-24       3       1 2019-10-26       1    <NA>      0
#> 123639: 2019-12-23 2019-12-19       2       2       <NA>       1    <NA>      3
#> 123640: 2019-10-03 2019-09-28       2       1 2019-10-03       1    <NA>      0
#> 123641: 2019-07-05 2019-07-02       2       2       <NA>       1    <NA>      7
#>          fecha_nto cer_def cbmte    fec_aju fm_fuerza     fm_unidad fm_grado
#>      1: 2014-09-21      NA       2019-08-02        NA  0.000000e+00         
#>      2: 1987-11-27      NA       2020-06-17        NA  0.000000e+00         
#>      3: 2013-09-24      NA       2019-09-02        NA  0.000000e+00         
#>      4: 2014-06-27      NA       2019-09-02        NA  0.000000e+00         
#>      5: 2012-10-13      NA       2019-09-04        NA  0.000000e+00         
#>     ---                                                                     
#> 123637: 2016-12-28      NA       2020-01-02        NA  0.000000e+00         
#> 123638: 1998-06-27      NA       2019-11-05         3 5.435637e-315      066
#> 123639: 2017-01-29      NA       2020-01-08        NA  0.000000e+00         
#> 123640: 1998-07-20      NA       2019-10-07        NA  0.000000e+00         
#> 123641: 2017-09-01      NA       2020-06-17        NA  0.000000e+00         
#>         confirmados estado_final_de_caso             nom_est_f_caso
#>      1:           1                    3 Confirmado por laboratorio
#>      2:           0                    2                   Probable
#>      3:           0                    2                   Probable
#>      4:           0                    2                   Probable
#>      5:           0                    2                   Probable
#>     ---                                                            
#> 123637:           1                    3 Confirmado por laboratorio
#> 123638:           1                    3 Confirmado por laboratorio
#> 123639:           1                    3 Confirmado por laboratorio
#> 123640:           0                    2                   Probable
#> 123641:           0                    2                   Probable
#>         departamento_ocurrencia municipio_ocurrencia departamento_residencia
#>      1:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      2:                    META              ACACIAS                    META
#>      3:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      4:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      5:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>     ---                                                                     
#> 123637:                 CAQUETA            FLORENCIA                 CAQUETA
#> 123638:            CUNDINAMARCA                 NILO                  BOGOTA
#> 123639:                   SUCRE            SINCELEJO                   SUCRE
#> 123640:                  TOLIMA                GUAMO                  TOLIMA
#> 123641:                  TOLIMA              ESPINAL                  TOLIMA
#>         municipio_residencia nombre_evento departamento_notificacion
#>      1:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      2:              ACACIAS        DENGUE                      META
#>      3:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      4:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      5:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>     ---                                                             
#> 123637:            FLORENCIA        DENGUE                   CAQUETA
#> 123638:               BOGOTA        DENGUE                    BOGOTA
#> 123639:            SINCELEJO        DENGUE                     SUCRE
#> 123640:                GUAMO        DENGUE                    TOLIMA
#> 123641:              ESPINAL        DENGUE                    TOLIMA
#>         municipio_notificacion
#>      1:                  PASTO
#>      2:                ACACIAS
#>      3:          MAGUI (PAYAN)
#>      4:          MAGUI (PAYAN)
#>      5:          MAGUI (PAYAN)
#>     ---                       
#> 123637:              FLORENCIA
#> 123638:                 BOGOTA
#> 123639:              SINCELEJO
#> 123640:                  GUAMO
#> 123641:                ESPINAL