Formatear fechas

Función que da un formato específico a una fecha

Usage

format_fecha(data_event, format_fecha = "%AAAA-%MM-%DD", nombres_col = c())

Arguments

data_event

Un data frame que contiene los datos de un evento o enfermedad

format_fecha

Un character (cadena de caracteres) que contiene el formato deseado de fecha

nombres_col

Un character (cadena de caracteres) que contiene los nombres de la columna a formatear

Value

Un data framecon los datos con las fechas formateadas

Examples

data_event <- import_data_event(2020, "DENGUE")
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
data_event <- format_fecha(data_event)
format_fecha(data_event = data_event,
             format_fecha = "%AAAA-%MM-%DD",
             nombres_col = c("ini_sin", "fec_hos"))
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
#>        CONSECUTIVE COD_EVE    FEC_NOT SEMANA  ANO       COD_PRE COD_SUB EDAD
#>     1:     7212061     210 2020-04-02     14 2020 3.756927e-314       1   13
#>     2:     7212037     210 2020-04-06     14 2020 1.351813e-314       1   35
#>     3:     7212059     210 2020-01-22      4 2020 2.040392e-314       1   29
#>     4:     7212047     210 2020-09-28     39 2020 2.481346e-314       1    6
#>     5:     7213565     210 2020-09-14     37 2020 1.259317e-314      80    4
#>    ---                                                                      
#> 76415:     7406806     210 2020-02-08      5 2020 3.754984e-314       2   76
#> 76416:     7406810     210 2020-02-24      7 2020 3.754948e-314      30   61
#> 76417:     7406790     210 2020-12-05     49 2020 2.052892e-314       1   43
#> 76418:     7406817     210 2020-11-03     45 2020 4.277176e-314       1   35
#> 76419:     7406816     210 2020-04-16     16 2020 3.770019e-314       1   75
#>        UNI_MED NACIONALID NOM_NACIONALIDAD SEXO COD_PAIS_O COD_DPTO_O COD_MUN_O
#>     1:       1        170         COLOMBIA    F        170         76        41
#>     2:       1        170         COLOMBIA    F        170         27       361
#>     3:       1        170         COLOMBIA    F        170         41       298
#>     4:       2        170         COLOMBIA    F        170         50       223
#>     5:       1        170         COLOMBIA    M        170         73       449
#>    ---                                                                         
#> 76415:       1        170         COLOMBIA    M        170         76         1
#> 76416:       1        170         COLOMBIA    M        170         76         1
#> 76417:       1        170         COLOMBIA    M        170         41       551
#> 76418:       1        170         COLOMBIA    M        170         86       571
#> 76419:       1        170         COLOMBIA    M        170         76       306
#>        AREA    LOCALIDAD CEN_POBLA      VEREDA        BAR_VER OCUPACION TIP_SS
#>     1:    1 ANSERMANUEVO  SIN DATO    SIN DATO   PUEBLO NUEVO      9997      S
#>     2:    1      ISTMINA  SIN DATO    SIN DATO          1E+40      9999      S
#>     3:    1     SIN DATO  SIN DATO    SIN DATO     SAN FELIPE      9996      S
#>     4:    1     SIN DATO  SIN DATO    SIN DATO        EL EDEN      9999      S
#>     5:    1       MELGAR  SIN DATO    SIN DATO     LA SULTANA      9999      P
#>    ---                                                                        
#> 76415:    1         CALI  SIN DATO    SIN DATO       SIN DATO      9995      P
#> 76416:    1     SIN DATO  SIN DATO    SIN DATO 0808 MUNICIPAL      9999      C
#> 76417:    1     SIN DATO  SIN DATO    SIN DATO       LAURELES      9622      C
#> 76418:    3     SIN DATO  SIN DATO SANTA LUCIA       SIN DATO      7549      S
#> 76419:    1     SIN DATO  SIN DATO    SIN DATO         CENTRO      7534      S
#>        COD_ASE PER_ETN GRU_POB NOM_GRUPO ESTRATO GP_DISCAPA GP_DESPLAZ
#>     1:  EPSS16       6      NA                 2          2          2
#>     2:  ESS076       5      NA                 1          2          2
#>     3:  CCF024       6      NA                NA          2          2
#>     4:  EPSS45       6      NA                 2          2          2
#>     5:  RES003       6      NA                 3          2          2
#>    ---                                                                
#> 76415:  RES001       6      NA                 3          2          2
#> 76416:  EPS018       6      NA                 3          2          2
#> 76417:  EPS005       6      NA                 1          2          2
#> 76418:  EPSI03       5      NA                 1          2          2
#> 76419:  ESS062       6      NA                 1          2          2
#>        GP_MIGRANT GP_CARCELA GP_GESTAN SEM_GES GP_INDIGEN GP_POBICFB GP_MAD_COM
#>     1:          2          2         2      NA          2          2          2
#>     2:          2          2         2      NA          2          2          2
#>     3:          2          2         2      NA          2          2          2
#>     4:          2          2         2      NA          2          2          2
#>     5:          2          2         2      NA          2          2          2
#>    ---                                                                         
#> 76415:          2          2         2      NA          2          2          2
#> 76416:          2          2         2      NA          2          2          2
#> 76417:          2          2         2      NA          2          2          2
#> 76418:          2          2         2      NA          2          2          2
#> 76419:          2          2         2      NA          2          2          2
#>        GP_DESMOVI GP_PSIQUIA GP_VIC_VIO GP_OTROS FUENTE COD_DPTO_R COD_MUN_R
#>     1:          2          2          2        1      1         76        41
#>     2:          2          2          2        1      1         27       361
#>     3:          2          2          2        1      1         41       298
#>     4:          2          2          2        1      1         50       223
#>     5:          2          2          2        1      1         73       449
#>    ---                                                                      
#> 76415:          2          2          2        1      1         76         1
#> 76416:          2          2          2        1      1         76         1
#> 76417:          2          2          2        1      1         41       551
#> 76418:          2          2          2        1      1         86       571
#> 76419:          2          2          2        1      1         76       306
#>        COD_DPTO_N COD_MUN_N    FEC_CON    INI_SIN TIP_CAS PAC_HOS    FEC_HOS
#>     1:         76     76041 2020-04-02 2020-03-29       2       2       <NA>
#>     2:         27     27361 2020-04-06 2020-04-03       2       2       <NA>
#>     3:         41     41298 2020-01-22 2020-01-19       2       2       <NA>
#>     4:         50     50223 2020-09-28 2020-09-26       2       1 2020-09-28
#>     5:         25     25488 2020-09-14 2020-09-11       2       2       <NA>
#>    ---                                                                      
#> 76415:         76     76001 2020-02-01 2020-01-26       3       2       <NA>
#> 76416:         76     76001 2020-02-19 2020-02-12       2       2       <NA>
#> 76417:         41     41551 2020-12-05 2020-12-04       2       2       <NA>
#> 76418:         86     86571 2020-11-03 2020-11-02       2       2       <NA>
#> 76419:         76     76306 2020-04-16 2020-04-14       2       2       <NA>
#>        CON_FIN FEC_DEF AJUSTE  FECHA_NTO CER_DEF CBMTE FEC_ARC_XL    FEC_AJU
#>     1:       1      NA      0 2006-06-20      NA    NA 2021-04-30 2020-04-04
#>     2:       1      NA      0 1984-10-28      NA    NA 2021-04-30 2020-04-06
#>     3:       1      NA      5 1990-02-28      NA    NA 2021-04-30 2020-09-04
#>     4:       1      NA      0 2020-03-27      NA    NA 2021-04-30 2020-10-05
#>     5:       1      NA      3 2016-06-15      NA    NA 2021-04-30 2020-10-11
#>    ---                                                                      
#> 76415:       1      NA      0 1943-09-19      NA    NA 2021-04-30 2020-02-09
#> 76416:       1      NA      0 1958-11-28      NA    NA 2021-04-30 2020-02-24
#> 76417:       1      NA      0 1977-02-19      NA    NA 2021-04-30 2020-12-18
#> 76418:       1      NA      0 1985-09-09      NA    NA 2021-04-30 2020-11-05
#> 76419:       1      NA      0 1944-09-09      NA    NA 2021-04-30 2020-04-27
#>        FM_FUERZA FM_UNIDAD FM_GRADO confirmado va_sispro est_f_caso
#>     1:        NA         0                   0         1          2
#>     2:        NA         0                   0         1          2
#>     3:        NA         0                   1         1          5
#>     4:        NA         0                   0         1          2
#>     5:        NA         0                   1         1          3
#>    ---                                                             
#> 76415:        NA         0                   1         1          3
#> 76416:        NA         0                   0         1          2
#> 76417:        NA         0                   0         1          2
#> 76418:        NA         0                   0         1          2
#> 76419:        NA         0                   0         1          2
#>                               nom_est_f_caso Nombre_evento
#>     1:                              Probable        DENGUE
#>     2:                              Probable        DENGUE
#>     3: Confirmado por Nexo Epidemiol\xf3gico        DENGUE
#>     4:                              Probable        DENGUE
#>     5:            Confirmado por laboratorio        DENGUE
#>    ---                                                    
#> 76415:            Confirmado por laboratorio        DENGUE
#> 76416:                              Probable        DENGUE
#> 76417:                              Probable        DENGUE
#> 76418:                              Probable        DENGUE
#> 76419:                              Probable        DENGUE
#>        Departamento_Notificacion Municipio_notificacion Departamento_ocurrencia
#>     1:                     VALLE           ANSERMANUEVO                   VALLE
#>     2:                     CHOCO                ITSMINA                   CHOCO
#>     3:                     HUILA                 GARZON                   HUILA
#>     4:                      META               CUBARRAL                    META
#>     5:              CUNDINAMARCA                   NILO                  TOLIMA
#>    ---                                                                         
#> 76415:                     VALLE                   CALI                   VALLE
#> 76416:                     VALLE                   CALI                   VALLE
#> 76417:                     HUILA               PITALITO                   HUILA
#> 76418:                  PUTUMAYO          PUERTO GUZMAN                PUTUMAYO
#> 76419:                     VALLE                GINEBRA                   VALLE
#>        Municipio_ocurrencia Departamento_residencia Municipio_residencia
#>     1:         ANSERMANUEVO                   VALLE         ANSERMANUEVO
#>     2:              ITSMINA                   CHOCO              ITSMINA
#>     3:               GARZON                   HUILA               GARZON
#>     4:             CUBARRAL                    META             CUBARRAL
#>     5:               MELGAR                  TOLIMA               MELGAR
#>    ---                                                                  
#> 76415:                 CALI                   VALLE                 CALI
#> 76416:                 CALI                   VALLE                 CALI
#> 76417:             PITALITO                   HUILA             PITALITO
#> 76418:        PUERTO GUZMAN                PUTUMAYO        PUERTO GUZMAN
#> 76419:              GINEBRA                   VALLE              GINEBRA