Filtrar por departamentos y municipios

Función que filtra los datos de una enfermedad o evento por departamentos y municipios

Uso

geo_filtro(data_event, dpto = NULL, mpio = NULL)

Argumentos

data_event

Un `data.frame` con los datos de una enfermedad o evento

dpto

Un `character` (cadena de caracteres) que contiene el nombre o código del departamento; valor por defecto `NULL`

mpio

Un `character` (cadena de caracteres) que contiene el nombre o código del municipio; su valor por defecto es `NULL`

Valor

Un `data.frame` con los datos filtrados con la enfermedad, departamentos y municipios seleccionados

Ejemplos

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
geo_filtro(data_event = data_limpia, dpto = "ANTIOQUIA")
#> # A tibble: 1,781 × 73
#>    consecutive cod_eve fec_not    semana ano   cod_pre    cod_sub  edad uni_med
#>    <chr>       <chr>   <date>     <chr>  <chr> <chr>      <chr>   <dbl>   <dbl>
#>  1 7233789     210     2020-07-07 27     2020  0508805568 07         27       1
#>  2 7233816     210     2020-06-02 22     2020  2300100487 02          8       1
#>  3 7233817     210     2020-07-11 28     2020  0579005513 01         13       1
#>  4 7233837     210     2020-02-23 08     2020  0566002277 01         15       1
#>  5 7233481     210     2020-10-20 41     2020  0526607137 01         31       1
#>  6 7233482     210     2020-03-13 11     2020  0504509099 35         15       1
#>  7 7233524     210     2020-11-18 46     2020  0557902201 02         36       1
#>  8 7233531     210     2020-08-17 33     2020  0561502129 01         27       1
#>  9 7233591     210     2020-05-09 18     2020  0539005090 01          5       1
#> 10 7233637     210     2020-09-10 37     2020  0561502129 01         46       1
#> # ℹ 1,771 more rows
#> # ℹ 64 more variables: nacionalidad <chr>, nombre_nacionalidad <chr>,
#> #   sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>, cod_mun_o <chr>,
#> #   area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>, per_etn <chr>,
#> #   gru_pob <lgl>, nom_grupo <chr>, estrato <chr>, gp_discapa <chr>,
#> #   gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>, gp_gestan <chr>,
#> #   sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, gp_mad_com <chr>, …
geo_filtro(data_event = data_limpia, dpto = "ANTIOQUIA", mpio = "ENVIGADO")
#> # A tibble: 23 × 73
#>    consecutive cod_eve fec_not    semana ano   cod_pre    cod_sub  edad uni_med
#>    <chr>       <chr>   <date>     <chr>  <chr> <chr>      <chr>   <dbl>   <dbl>
#>  1 7248466     210     2020-01-10 02     2020  0500100302 01         29       1
#>  2 7226063     210     2020-07-17 29     2020  0526604910 01          9       1
#>  3 7320529     210     2020-01-17 03     2020  0526605568 10         23       1
#>  4 7298493     210     2020-08-17 33     2020  0500111746 01         22       1
#>  5 7294504     210     2020-04-25 16     2020  0526605568 10         32       1
#>  6 7281492     210     2020-07-03 27     2020  0526604910 01         30       1
#>  7 7281986     210     2020-07-06 27     2020  0526604910 01         18       1
#>  8 7375007     210     2020-10-19 42     2020  0500105482 01         12       1
#>  9 7355822     210     2020-04-08 14     2020  0526605568 10         59       1
#> 10 7354610     210     2020-02-12 06     2020  0526605568 10         45       1
#> # ℹ 13 more rows
#> # ℹ 64 more variables: nacionalidad <chr>, nombre_nacionalidad <chr>,
#> #   sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>, cod_mun_o <chr>,
#> #   area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>, per_etn <chr>,
#> #   gru_pob <lgl>, nom_grupo <chr>, estrato <chr>, gp_discapa <chr>,
#> #   gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>, gp_gestan <chr>,
#> #   sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, gp_mad_com <chr>, …