Importar los datos de una enfermedad o evento por año desde los microdatos de SIVIGILA

Función que obtiene los datos de una enfermedad por año desde los microdatos de SIVIGILA

Usage

import_data_event(year, nombre_event, cache = TRUE)

Arguments

year

Un numeric (numerico) con el año deseado para la descarga de los datos

nombre_event

Un character (cadena de caracteres) con el nombre de la enfermedad o evento

cache

Un boolean (TRUE o FALSE) que indica si los datos descargados deben ser almacenados en caché; su valor por defecto es TRUE

Value

Un `data.frame` con los datos de la enfermedad o evento seleccionado por año desde los microdatos de SIVIGILA

Examples

import_data_event(2018, "DENGUE")
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
#>        CONSECUTIVE COD_EVE    FEC_NOT SEMANA  ANO       COD_PRE COD_SUB EDAD
#>     1:     5622015     210 2018-10-07     40 2018 5.435248e-315      18   40
#>     2:     5627187     210 2018-07-17     28 2018 4.496047e-314       1   42
#>     3:     5592910     210 2018-01-12      2 2018 4.496047e-314       1   24
#>     4:     5583361     210 2018-11-30     47 2018 5.435261e-315       1   18
#>     5:     5623291     210 2018-12-10     49 2018 5.435335e-315       4   64
#>    ---                                                                      
#> 43648:     5600170     210 2018-05-05     17 2018 4.693715e-314      80    3
#> 43649:     5618830     210 2018-04-21     16 2018 4.693673e-314       1   43
#> 43650:     5598354     210 2018-11-19     45 2018 4.693673e-314       1   19
#> 43651:     5592318     210 2018-12-29     50 2018 9.882207e-315      80   25
#> 43652:     5611985     210 2018-06-14     24 2018 4.693673e-314       1   13
#>        UNI_MED SEXO COD_PAIS_O COD_DPTO_O COD_MUN_O AREA LOCALIDAD CEN_POBLA
#>     1:       1    F        356          1       356    1  SIN DATO  SIN DATO
#>     2:       1    M        604          1       604    2  SIN DATO  SIN DATO
#>     3:       1    M        604          1       604    2  SIN DATO  SIN DATO
#>     4:       1    F        484          1       484    1  SIN DATO  SIN DATO
#>     5:       1    M        484          1       484    2  SIN DATO  SIN DATO
#>    ---                                                                      
#> 43648:       1    F        170         95         1    1  SIN DATO  SIN DATO
#> 43649:       1    M        170         95         1    2  SIN DATO  SIN DATO
#> 43650:       1    M        170         95         1    3  SIN DATO  SIN DATO
#> 43651:       1    M        170         95         1    3  SIN DATO  SIN DATO
#> 43652:       1    M        170         95       200    1  SIN DATO  SIN DATO
#>        VEREDA BAR_VER OCUPACION TIP_SS COD_ASE PER_ETN GRU_POB GP_DISCAPA
#>     1:     NA      NA      4222      C  EPS008       6      NA          2
#>     2:     NA      NA      9622      S  EPSS41       6      NA          2
#>     3:     NA      NA      9997      S  ESS062       1      NA          2
#>     4:     NA      NA      9997      C  EPS010       6      NA          2
#>     5:     NA      NA      9999      C  EPS016       6      NA          2
#>    ---                                                                   
#> 43648:     NA      NA      9999      C  CCFC24       6      NA          2
#> 43649:     NA      NA      4414      C  EPS037       6      NA          2
#> 43650:     NA      NA      9994      P  RES003       6      NA          2
#> 43651:     NA      NA      9994      S  EPSS33       6      NA          2
#> 43652:     NA      NA      9997      S  EPSS41       1      NA          2
#>        GP_DESPLAZ GP_MIGRANT GP_CARCELA GP_GESTAN GP_INDIGEN GP_POBICFB
#>     1:          2          2          2         2          2          2
#>     2:          2          2          2         2          2          2
#>     3:          2          2          2         2          2          2
#>     4:          2          2          2         2          2          2
#>     5:          2          2          2         2          2          2
#>    ---                                                                 
#> 43648:          2          2          2         2          2          2
#> 43649:          2          2          2         2          2          2
#> 43650:          2          2          2         2          2          2
#> 43651:          2          2          2         2          2          2
#> 43652:          2          2          2         2          2          2
#>        GP_MAD_COM GP_DESMOVI GP_PSIQUIA GP_VIC_VIO GP_OTROS COD_DPTO_R
#>     1:          2          2          2          2        1         11
#>     2:          2          2          2          2        1         91
#>     3:          2          2          2          2        1         19
#>     4:          2          2          2          2        1         11
#>     5:          2          2          2          2        1         11
#>    ---                                                                
#> 43648:          2          2          2          2        1         95
#> 43649:          2          2          2          2        1          0
#> 43650:          2          2          2          2        1         95
#> 43651:          2          2          2          2        1         20
#> 43652:          2          2          2          2        1         95
#>        COD_MUN_R COD_DPTO_N COD_MUN_N    FEC_CON    INI_SIN TIP_CAS PAC_HOS
#>     1:         1         11     11001 2018-10-07 2018-10-03       3       1
#>     2:         1         91     91001 2018-07-17 2018-07-14       2       1
#>     3:         0         91     91001 2018-01-12 2018-01-08       2       2
#>     4:         1         11     11001 2018-11-29 2018-11-22       3       2
#>     5:         1         11     11001 2018-12-10 2018-12-05       3       2
#>    ---                                                                     
#> 43648:         1         95     95001 2018-05-03 2018-04-28       3       2
#> 43649:         0         95     95001 2018-04-18 2018-04-16       2       2
#> 43650:        15         95     95001 2018-11-17 2018-11-10       2       1
#> 43651:         1         20     20001 2018-12-28 2018-12-12       2       2
#> 43652:       200         95     95001 2018-06-14 2018-06-11       2       2
#>           FEC_HOS CON_FIN FEC_DEF AJUSTE  FECHA_NTO CER_DEF CBMTE FEC_ARC_XL
#>     1: 2018-10-07       1      NA      0 1978-07-27      NA    NA 2019-05-08
#>     2: 2018-07-17       1      NA      0 1976-01-03      NA    NA 2019-05-08
#>     3:       <NA>       1      NA      7 1993-08-20      NA    NA 2019-05-08
#>     4:       <NA>       1      NA      0 2000-06-29      NA    NA 2019-05-08
#>     5:       <NA>       1      NA      0 1954-06-09      NA    NA 2019-05-08
#>    ---                                                                      
#> 43648:       <NA>       1      NA      7 2014-06-16      NA    NA 2019-05-08
#> 43649:       <NA>       1      NA      7 1974-09-28      NA    NA 2019-05-08
#> 43650: 2018-11-17       1      NA      0 1998-12-03      NA    NA 2019-05-08
#> 43651:       <NA>       1      NA      7 1993-08-01      NA    NA 2019-05-08
#> 43652:       <NA>       1      NA      3 2004-12-05      NA    NA 2019-05-08
#>           FEC_AJU FM_FUERZA     FM_UNIDAD FM_GRADO VERSION confirmados
#>     1: 2018-10-11        NA  0.000000e+00               NA           1
#>     2: 2018-07-23        NA  0.000000e+00               NA           0
#>     3: 2019-03-13        NA  0.000000e+00               NA           0
#>     4: 2018-11-30        NA  0.000000e+00               NA           1
#>     5: 2018-12-12        NA  0.000000e+00               NA           1
#>    ---                                                                
#> 43648: 2019-03-13        NA  0.000000e+00               NA           1
#> 43649: 2019-03-13        NA  0.000000e+00               NA           0
#> 43650: 2018-11-25        NA  0.000000e+00               NA           0
#> 43651: 2019-03-13         3 1.492335e-314       64      NA           0
#> 43652: 2018-07-16        NA  0.000000e+00               NA           1
#>        est_f_caso Evento       estado_final_de_caso Departanento_ocurrencia
#>     1:          3 DENGUE Confirmado por laboratorio                EXTERIOR
#>     2:          2 DENGUE                   Probable                EXTERIOR
#>     3:          2 DENGUE                   Probable                EXTERIOR
#>     4:          3 DENGUE Confirmado por laboratorio                EXTERIOR
#>     5:          3 DENGUE Confirmado por laboratorio                EXTERIOR
#>    ---                                                                     
#> 43648:          3 DENGUE Confirmado por laboratorio                GUAVIARE
#> 43649:          2 DENGUE                   Probable                GUAVIARE
#> 43650:          2 DENGUE                   Probable                GUAVIARE
#> 43651:          2 DENGUE                   Probable                GUAVIARE
#> 43652:          3 DENGUE Confirmado por laboratorio                GUAVIARE
#>         Municipio_ocurrencia Departamento_residencia
#>     1:        EXTERIOR_INDIA                  BOGOTA
#>     2:         EXTERIOR_PERU                AMAZONAS
#>     3:         EXTERIOR_PERU                   CAUCA
#>     4:    EXTERIOR_M\xc9XICO                  BOGOTA
#>     5:    EXTERIOR_M\xc9XICO                  BOGOTA
#>    ---                                              
#> 43648: SAN JOSE DEL GUAVIARE                GUAVIARE
#> 43649: SAN JOSE DEL GUAVIARE PROCEDENCIA DESCONOCIDA
#> 43650: SAN JOSE DEL GUAVIARE                GUAVIARE
#> 43651: SAN JOSE DEL GUAVIARE                   CESAR
#> 43652:            MIRAFLORES                GUAVIARE
#>                  Municipio_residencia
#>     1:                         BOGOTA
#>     2:                        LETICIA
#>     3: * CAUCA. MUNICIPIO DESCONOCIDO
#>     4:                         BOGOTA
#>     5:                         BOGOTA
#>    ---                               
#> 43648:          SAN JOSE DEL GUAVIARE
#> 43649:      * PROCEDENCIA DESCONOCIDA
#> 43650:                        CALAMAR
#> 43651:                     VALLEDUPAR
#> 43652:                     MIRAFLORES