Limpiar códigos de departamento de los datos de una enfermedad o evento

Función que limpia los códigos de departamento de los datos de una enfermedad o evento

limpiar_cods_dpto(data_event, col_cods_data, geo_data, col_geo_cods)

Argumentos

data_event

Un `data.frame` que contiene los datos de una enfermedad o evento

col_cods_data

Un `character` (cadena de caracteres) que contiene el nombre de la columna en los datos de una enfermedad o evento que contiene los códigos de departamento

geo_data

Un `data.frame` con los datos geográficos que incluyen los códigos de departamento

col_geo_cods

Un `character` (cadena de caracteres) con el nombre de la columna en los datos geográficos que contiene los códigos de departamento

Valor

Un `data.frame` con los códigos de departamento limpios de los datos de una enfermedad o evento

Examples

data(dengue2020)
geo_codes <- import_geo_cods()
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
limpiar_cods_dpto(data_event = data_limpia,
                  col_cods_data = "cod_dpto_o",
                  geo_data = geo_codes,
                  col_geo_cods = "codigo_departamento")
#> # A tibble: 63,742 × 73
#>    consecutive cod_eve fec_not    semana ano   cod_pre    cod_sub  edad uni_med
#>    <chr>       <chr>   <date>     <chr>  <chr> <chr>      <chr>   <dbl>   <dbl>
#>  1 7232769     210     2020-02-27 09     2020  6800104268 01         15       1
#>  2 7232770     210     2020-02-08 05     2020  5011000634 11          4       1
#>  3 7232771     210     2020-03-04 09     2020  7600103957 11         15       1
#>  4 7232772     210     2020-01-21 03     2020  2000100438 01         10       1
#>  5 7232773     210     2020-05-19 20     2020  8500100001 03         10       1
#>  6 7232774     210     2020-04-13 15     2020  7689504656 01          9       1
#>  7 7232775     210     2020-03-11 11     2020  7603601707 01          2       1
#>  8 7232776     210     2020-05-05 18     2020  7600103359 01          4       1
#>  9 7232777     210     2020-07-29 31     2020  7689504656 01          2       1
#> 10 7232778     210     2020-06-20 25     2020  7624804722 01          5       1
#> # ℹ 63,732 more rows
#> # ℹ 64 more variables: nacionalidad <chr>, nombre_nacionalidad <chr>,
#> #   sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>, cod_mun_o <chr>,
#> #   area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>, per_etn <chr>,
#> #   gru_pob <lgl>, nom_grupo <chr>, estrato <chr>, gp_discapa <chr>,
#> #   gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>, gp_gestan <chr>,
#> #   sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, gp_mad_com <chr>, …