Limpiar datos de SIVIGILA

Función que limpia los datos seleccionados de una enfermedad o evento de la fuente de SIVIGILA

Usage

limpiar_data_sivigila(data_event, year)

Arguments

data_event

Un data frame que contiene los datos de una enfermedad o evento

year

Un numeric (numerico) que contiene el año de los datos de una enfermedad o evento

Value

Un data framecon los datos limpios de la enfermedad o evento

Examples

year <- 2019
data_event <- import_data_event(year, "DENGUE")
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
limpiar_data_sivigila(data_event = data_event, year = year)
#>         cod_eve    fec_not semana  ano       cod_pre cod_sub edad uni_med
#>      1:     210 2019-05-19     20 2019 2.569191e-314       1    4       1
#>      2:     210 2019-09-17     37 2019 2.470625e-314       1   31       1
#>      3:     210 2019-08-30     35 2019 2.590238e-314       1    5       1
#>      4:     210 2019-08-30     35 2019 2.590238e-314       1    5       1
#>      5:     210 2019-09-02     35 2019 2.590238e-314       1    6       1
#>     ---                                                                  
#> 121505:     210 2019-12-23     51 2019 8.893713e-315       1    2       1
#> 121506:     210 2019-11-01     43 2019 5.435637e-315      80   21       1
#> 121507:     210 2019-12-23     51 2019 3.458509e-314       1    2       1
#> 121508:     210 2019-10-03     39 2019 3.622440e-314       1   21       1
#> 121509:     210 2019-07-05     27 2019 3.619921e-314       1    1       1
#>         nacionalidad nombre_nacionalidad sexo cod_pais_o cod_dpto_o cod_mun_o
#>      1:          170            COLOMBIA    F        170         52       427
#>      2:          170            COLOMBIA    F        170         50       006
#>      3:          170            COLOMBIA    F        170         52       427
#>      4:          170            COLOMBIA    M        170         52       427
#>      5:          170            COLOMBIA    F        170         52       427
#>     ---                                                                      
#> 121505:          170            COLOMBIA    F        170         18       001
#> 121506:          170            COLOMBIA    F        170         25       488
#> 121507:          170            COLOMBIA    M        170         70       001
#> 121508:          170            COLOMBIA    F        170         73       319
#> 121509:          170            COLOMBIA    M        170         73       268
#>         area localidad       cen_pobla                       vereda
#>      1:    2  SIN DATO DOCE DE OCTUBRE                     SIN DATO
#>      2:    1   ACACIAS        SIN DATO                     SIN DATO
#>      3:    1     MAGUI        SIN DATO                     SIN DATO
#>      4:    1     MAGUI        SIN DATO                     SIN DATO
#>      5:    3  SIN DATO        SIN DATO                       PLAYON
#>     ---                                                            
#> 121505:    1 FLORENCIA        SIN DATO                     SIN DATO
#> 121506:    1  SIN DATO        SIN DATO                     SIN DATO
#> 121507:    1  SIN DATO        SIN DATO                     SIN DATO
#> 121508:    1     GUAMO        SIN DATO                     SIN DATO
#> 121509:    3  SIN DATO        SIN DATO AGUA BLANCA SEC. DE LA DULCE
#>                        bar_ver ocupacion tip_ss cod_ase per_etn gru_pob
#>      1:               SIN DATO      9997      S  ESS062       5      NA
#>      2: PORTALES DE SAN CARLOS      9996      C  EPS013       6      NA
#>      3:                PLAYITA      9997      S  ESS118       5      NA
#>      4:             EL PARAISO      9999      S  ESS062       5      NA
#>      5:               SIN DATO      9997      S  ESS062       5      NA
#>     ---                                                                
#> 121505:                 CENTRO      9999      S  EPSS44       6      NA
#> 121506:               SIN DATO      9997      P  RES003       6      NA
#> 121507:        SIN INFORMACION      9999      C  EPS002       6      NA
#> 121508:             LIBERTADOR      9997      C  EPS044       6      NA
#> 121509:               SIN DATO      9999      C  EPS044       6      NA
#>         nom_grupo estrato gp_discapa gp_desplaz gp_migrant gp_carcela gp_gestan
#>      1:                 1          2          2          2          2         2
#>      2:                 3          2          2          2          2         2
#>      3:                 1          2          2          2          2         2
#>      4:                 1          2          2          2          2         2
#>      5:                 1          2          2          2          2         2
#>     ---                                                                        
#> 121505:                NA          2          2          2          2         2
#> 121506:                 3          2          2          2          2         2
#> 121507:                NA          2          2          2          2         2
#> 121508:                 2          2          2          2          2         2
#> 121509:                NA          2          2          2          2         2
#>         sem_ges gp_indigen gp_pobicfb gp_mad_com gp_desmovi gp_psiquia
#>      1:      NA          2          2          2          2          2
#>      2:      NA          2          2          2          2          2
#>      3:      NA          2          2          2          2          2
#>      4:      NA          2          2          2          2          2
#>      5:      NA          2          2          2          2          2
#>     ---                                                               
#> 121505:      NA          2          2          2          2          2
#> 121506:      NA          2          2          2          2          2
#> 121507:      NA          2          2          2          2          2
#> 121508:      NA          2          2          2          2          2
#> 121509:      NA          2          2          2          2          2
#>         gp_vic_vio gp_otros fuente cod_dpto_r cod_mun_r cod_dpto_n cod_mun_n
#>      1:          2        1      1         52       427         52     52001
#>      2:          2        1      1         50       006         50     50006
#>      3:          2        1      1         52       427         52     52427
#>      4:          2        1      1         52       427         52     52427
#>      5:          2        1      1         52       427         52     52427
#>     ---                                                                     
#> 121505:          2        1      1         18       001         18     18001
#> 121506:          2        1      1         11       001         11     11001
#> 121507:          2        1      1         70       001         70     70001
#> 121508:          2        1      1         73       319         73     73319
#> 121509:          2        1      1         73       268         73     73268
#>            fec_con    ini_sin tip_cas pac_hos    fec_hos con_fin fec_def ajuste
#>      1: 2019-05-18 2019-05-13       2       1 2019-05-18       1    <NA>      3
#>      2: 2019-09-17 2019-09-14       2       2       <NA>       1    <NA>      7
#>      3: 2019-08-30 2019-08-28       2       2       <NA>       1    <NA>      0
#>      4: 2019-08-30 2019-08-27       2       2       <NA>       1    <NA>      0
#>      5: 2019-09-02 2019-08-31       2       2       <NA>       1    <NA>      0
#>     ---                                                                        
#> 121505: 2019-12-21 2019-12-17       2       1 2019-12-21       1    <NA>      3
#> 121506: 2019-10-26 2019-10-24       3       1 2019-10-26       1    <NA>      0
#> 121507: 2019-12-23 2019-12-19       2       2       <NA>       1    <NA>      3
#> 121508: 2019-10-03 2019-09-28       2       1 2019-10-03       1    <NA>      0
#> 121509: 2019-07-05 2019-07-02       2       2       <NA>       1    <NA>      7
#>          fecha_nto cer_def cbmte    fec_aju fm_fuerza     fm_unidad fm_grado
#>      1: 2014-09-21      NA       2019-08-02        NA  0.000000e+00         
#>      2: 1987-11-27      NA       2020-06-17        NA  0.000000e+00         
#>      3: 2013-09-24      NA       2019-09-02        NA  0.000000e+00         
#>      4: 2014-06-27      NA       2019-09-02        NA  0.000000e+00         
#>      5: 2012-10-13      NA       2019-09-04        NA  0.000000e+00         
#>     ---                                                                     
#> 121505: 2016-12-28      NA       2020-01-02        NA  0.000000e+00         
#> 121506: 1998-06-27      NA       2019-11-05         3 5.435637e-315      066
#> 121507: 2017-01-29      NA       2020-01-08        NA  0.000000e+00         
#> 121508: 1998-07-20      NA       2019-10-07        NA  0.000000e+00         
#> 121509: 2017-09-01      NA       2020-06-17        NA  0.000000e+00         
#>         confirmados estado_final_de_caso             nom_est_f_caso
#>      1:           1                    3 Confirmado por laboratorio
#>      2:           0                    2                   Probable
#>      3:           0                    2                   Probable
#>      4:           0                    2                   Probable
#>      5:           0                    2                   Probable
#>     ---                                                            
#> 121505:           1                    3 Confirmado por laboratorio
#> 121506:           1                    3 Confirmado por laboratorio
#> 121507:           1                    3 Confirmado por laboratorio
#> 121508:           0                    2                   Probable
#> 121509:           0                    2                   Probable
#>         departamento_ocurrencia municipio_ocurrencia departamento_residencia
#>      1:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      2:                    META              ACACIAS                    META
#>      3:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      4:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>      5:               NARI\xd1O        MAGUI (PAYAN)               NARI\xd1O
#>     ---                                                                     
#> 121505:                 CAQUETA            FLORENCIA                 CAQUETA
#> 121506:            CUNDINAMARCA                 NILO                  BOGOTA
#> 121507:                   SUCRE            SINCELEJO                   SUCRE
#> 121508:                  TOLIMA                GUAMO                  TOLIMA
#> 121509:                  TOLIMA              ESPINAL                  TOLIMA
#>         municipio_residencia nombre_evento departamento_notificacion
#>      1:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      2:              ACACIAS        DENGUE                      META
#>      3:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      4:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>      5:        MAGUI (PAYAN)        DENGUE                 NARI\xd1O
#>     ---                                                             
#> 121505:            FLORENCIA        DENGUE                   CAQUETA
#> 121506:               BOGOTA        DENGUE                    BOGOTA
#> 121507:            SINCELEJO        DENGUE                     SUCRE
#> 121508:                GUAMO        DENGUE                    TOLIMA
#> 121509:              ESPINAL        DENGUE                    TOLIMA
#>         municipio_notificacion
#>      1:                  PASTO
#>      2:                ACACIAS
#>      3:          MAGUI (PAYAN)
#>      4:          MAGUI (PAYAN)
#>      5:          MAGUI (PAYAN)
#>     ---                       
#> 121505:              FLORENCIA
#> 121506:                 BOGOTA
#> 121507:              SINCELEJO
#> 121508:                  GUAMO
#> 121509:                ESPINAL