Limpiar las etiquetas del encabezado de los datos de una enfermedad

Función que limpia las etiquetas del encabezado de los datos de una enfermedad o evento

limpiar_encabezado(data_event)

Argumentos

data_event

Un `data.frame` que contiene los datos de una enfermedad o evento

Valor

Un `data.frame` con las etiquetas del encabezado formateadas con guión bajo (_)

Examples

data(dengue2020)
limpiar_encabezado(data_event = dengue2020)
#> # A tibble: 65,535 × 74
#>    particion      consecutive cod_eve fec_not semana ano   cod_pre cod_sub edad 
#>    <chr>          <chr>       <chr>   <chr>   <chr>  <chr> <chr>   <chr>   <chr>
#>  1 Datos_2020_210 7232769     210     2020-0… 09     2020  680010… 01      15   
#>  2 Datos_2020_210 7232770     210     2020-0… 05     2020  501100… 11      4    
#>  3 Datos_2020_210 7232771     210     2020-0… 09     2020  760010… 11      15   
#>  4 Datos_2020_210 7232772     210     2020-0… 03     2020  200010… 01      10   
#>  5 Datos_2020_210 7232773     210     2020-0… 20     2020  850010… 03      10   
#>  6 Datos_2020_210 7232774     210     2020-0… 15     2020  768950… 01      9    
#>  7 Datos_2020_210 7232775     210     2020-0… 11     2020  760360… 01      2    
#>  8 Datos_2020_210 7232776     210     2020-0… 18     2020  760010… 01      4    
#>  9 Datos_2020_210 7232777     210     2020-0… 31     2020  768950… 01      2    
#> 10 Datos_2020_210 7232778     210     2020-0… 25     2020  762480… 01      5    
#> # ℹ 65,525 more rows
#> # ℹ 65 more variables: uni_med <chr>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …