Obtener la población para efectuar el cálculo de la incidencia

Función que obtiene la población a riesgo de un evento o enfermedad o las proyecciones poblacionales DANE desde el año 2005 hasta el 2035. Si no hay población a riesgo disponible del evento o enfermedad para el año seleccionado se obtendrá las proyecciones poblacionales DANE y se mostrarán mensajes de advertencia al usuario dependendiendo del tipo de población obtenida

Uso

obtener_pob_incidencia(data_incidencia = NULL, poblacion, event, year)

Argumentos

data_incidencia

Un `data.frame` que contiene la población a riesgo o las proyecciones poblaciones DANE. Si este parámetro está vacío importará la población a riesgo o las proyecciones dependiendo de la disponibilidad de la información y las condiciones del evento o engermedad; su valor por defecto es `NULL`

poblacion

Un `character` (cadena de caracteres) con el tipo de población que se desea obtener. Indica si se desea obtener la población a riesgo del evento `"riesgo"` o las proyecciones poblacionales DANE `"proyecciones"`

event

Un `character` (cadena de caracteres) o un `numeric` (numerico) con el nombre o código de la enfermedad o evento. Es obligatorio para obtener la población a riesgo

year

Un `numeric` (numerico) con el año deseado de la población a riesgo. Es obligatorio para obtener la población a riesgo

Valor

Un `data.frame` con la población a riesgo o las proyecciones poblacionaldes DANE