Eliminar valores NIN (NA, Infinito, NaN)

Función que elimina filas si los valores de la columna seleccionada incluyen NA, Infinito o NaN

Uso

remove_val_nin(data_event, nomb_col)

Argumentos

data_event

Un `data.frame` que contiene los datos de una enfermedad o evento

nomb_col

Un `character` (cadena de caracteres) que contiene el nombre de la columna en los datos de una enfermedad o evento a evaluar

Valor

Los datos limpios sin valores NA, Infinito o NaN

Ejemplos

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
remove_val_nin(data_event = data_limpia, nomb_col = "edad")
#> # A tibble: 62,124 × 74
#>    particion   consecutive cod_eve fec_not    semana ano   cod_pre cod_sub  edad
#>    <chr>       <chr>       <chr>   <date>     <chr>  <chr> <chr>   <chr>   <dbl>
#>  1 Datos_2020… 7232769     210     2020-02-27 09     2020  680010… 01         15
#>  2 Datos_2020… 7232771     210     2020-03-04 09     2020  760010… 11         15
#>  3 Datos_2020… 7232772     210     2020-01-21 03     2020  200010… 01         10
#>  4 Datos_2020… 7232773     210     2020-05-19 20     2020  850010… 03         10
#>  5 Datos_2020… 7232774     210     2020-04-13 15     2020  768950… 01          9
#>  6 Datos_2020… 7232775     210     2020-03-11 11     2020  760360… 01          2
#>  7 Datos_2020… 7232776     210     2020-05-05 18     2020  760010… 01          4
#>  8 Datos_2020… 7232777     210     2020-07-29 31     2020  768950… 01          2
#>  9 Datos_2020… 7232778     210     2020-06-20 25     2020  762480… 01          5
#> 10 Datos_2020… 7232779     210     2020-01-27 04     2020  541720… 01          5
#> # ℹ 62,114 more rows
#> # ℹ 65 more variables: uni_med <dbl>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …