sivirep: Data Wrangling and Automated Reports from 'SIVIGILA' Source

Data wrangling, pre-processing, and generating automated reports from Colombia's epidemiological surveillance system, 'SIVIGILA' https://portalsivigila.ins.gov.co/. It provides a customizable R Markdown template for analysis and automatic generation of epidemiological reports that can be adapted to local, regional, and national contexts. This tool offers a standardized and reproducible workflow that helps to reduce manual labor and potential errors in report generation, improving their efficiency and consistency.

Ver también

Autor-a

Maintainer: Geraldine Gómez-Millán geralidine.gomez@javeriana.edu.co (ORCID) [contributor]

Authors:

Zulma M. Cucunubá zulma.cucunuba@javeriana.edu.co (ORCID) [contributor]

Jennifer A. Mendez-Romero jenniferk2@hotmail.com (ORCID) [contributor]

Claudia Huguett-Aragón chuguett@hotmail.com (ORCID) [contributor]

Other contributors:

Hugo Gruson hugo.gruson+R@normalesup.org (ORCID) [contributor]

Juanita Romero-Garcés juanita.romerog@javeriana.edu.co (ORCID) [contributor]

Jaime Pavlich-Mariscal jpavlich@javeriana.edu.co (ORCID) [contributor]

Laura Gómez Bermeo gomezblaura@javeriana.edu.co (ORCID) [contributor]

Andrés Moreno ad.morenob@javeriana.edu.co (ORCID) [contributor]

Miguel Gámez [contributor]

Johan Calderón [contributor]

Lady Flórez-Tapiero [contributor]

Verónica Tangarife-Arredondo [contributor]

Gerard Alarcon [contributor]

International Development Research Center (IDRC) [funder]

Pontificia Universidad Javeriana [copyright holder]