Agrupar por la clasificación inicial del caso

Función que agrupa los casos por la clasificación inicial del caso

Uso

agrupar_tipo_caso(data_event, cols_tipo = "tip_cas")

Argumentos

data_event

Un `data.frame` que contiene los datos de la enfermedad o evento

cols_tipo

Un `character` (cadena de caracteres) con el nombre de las columna(s) que contiene la clasificación inicial del caso en los datos de la enfermedad o evento; su valor por defecto es `"tip_cas"`

Valor

Un `data.frame` con los datos de la enfermedad o evento agrupados por la clasificación inicial del caso y/u otras variables como los años

Ejemplos

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
agrupar_tipo_caso(data_event = data_limpia,
                  cols_tipo = "tip_cas")
#> # A tibble: 3 × 6
#>   tip_cas cod_eve nombre_evento ano   casos nombre_tip_cas
#>   <chr>   <chr>   <chr>         <chr> <int> <named list>  
#> 1 2       210     DENGUE        2020     36 <chr [1]>     
#> 2 3       210     DENGUE        2020      8 <chr [1]>     
#> 3 5       210     DENGUE        2020      1 <chr [1]>