Filtrar por departamentos y municipios

Función que filtra los datos de una enfermedad o evento por departamentos y municipios

Uso

geo_filtro(data_event, dpto = NULL, mpio = NULL)

Argumentos

data_event

Un `data.frame` con los datos de una enfermedad o evento

dpto

Un `character` (cadena de caracteres) o `numeric` (numérico) que contiene el nombre o código del departamento; valor por defecto `NULL`

mpio

Un `character` (cadena de caracteres) o `numeric` (numérico) que contiene el nombre o código del municipio; su valor por defecto es `NULL`

Valor

Un `data.frame` con los datos filtrados con la enfermedad, departamentos y municipios seleccionados

Ejemplos

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
geo_filtro(data_event = data_limpia, dpto = "ANTIOQUIA")
#> # A tibble: 1,710 × 74
#>    particion   consecutive cod_eve fec_not    semana ano   cod_pre cod_sub  edad
#>    <chr>       <chr>       <chr>   <date>     <chr>  <chr> <chr>   <chr>   <dbl>
#>  1 Datos_2020… 7233789     210     2020-07-07 27     2020  050880… 07         27
#>  2 Datos_2020… 7233816     210     2020-06-02 22     2020  230010… 02          8
#>  3 Datos_2020… 7233817     210     2020-07-11 28     2020  057900… 01         13
#>  4 Datos_2020… 7233837     210     2020-02-23 08     2020  056600… 01         15
#>  5 Datos_2020… 7233481     210     2020-10-20 41     2020  052660… 01         31
#>  6 Datos_2020… 7233482     210     2020-03-13 11     2020  050450… 35         15
#>  7 Datos_2020… 7233524     210     2020-11-18 46     2020  055790… 02         36
#>  8 Datos_2020… 7233531     210     2020-08-17 33     2020  056150… 01         27
#>  9 Datos_2020… 7233591     210     2020-05-09 18     2020  053900… 01          5
#> 10 Datos_2020… 7233682     210     2020-12-20 50     2020  056150… 01          5
#> # ℹ 1,700 more rows
#> # ℹ 65 more variables: uni_med <dbl>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …
geo_filtro(data_event = data_limpia, dpto = "ANTIOQUIA", mpio = "MEDELLIN")
#> # A tibble: 504 × 74
#>    particion   consecutive cod_eve fec_not    semana ano   cod_pre cod_sub  edad
#>    <chr>       <chr>       <chr>   <date>     <chr>  <chr> <chr>   <chr>   <dbl>
#>  1 Datos_2020… 7233481     210     2020-10-20 41     2020  052660… 01         31
#>  2 Datos_2020… 7233755     210     2020-03-03 10     2020  050010… 08         15
#>  3 Datos_2020… 7345121     210     2020-09-10 36     2020  050010… 10         50
#>  4 Datos_2020… 7345145     210     2020-04-04 14     2020  050011… 07         33
#>  5 Datos_2020… 7234649     210     2020-02-06 05     2020  050010… 01         57
#>  6 Datos_2020… 7317731     210     2020-10-03 40     2020  050010… 01         23
#>  7 Datos_2020… 7317853     210     2020-01-22 03     2020  110013… 02          9
#>  8 Datos_2020… 7249163     210     2020-09-23 38     2020  052660… 03         66
#>  9 Datos_2020… 7249172     210     2020-01-18 02     2020  050010… 17         66
#> 10 Datos_2020… 7314789     210     2020-01-08 02     2020  050011… 01         27
#> # ℹ 494 more rows
#> # ℹ 65 more variables: uni_med <dbl>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …
geo_filtro(data_event = data_limpia, dpto = "05")
#> # A tibble: 1,710 × 74
#>    particion   consecutive cod_eve fec_not    semana ano   cod_pre cod_sub  edad
#>    <chr>       <chr>       <chr>   <date>     <chr>  <chr> <chr>   <chr>   <dbl>
#>  1 Datos_2020… 7233789     210     2020-07-07 27     2020  050880… 07         27
#>  2 Datos_2020… 7233816     210     2020-06-02 22     2020  230010… 02          8
#>  3 Datos_2020… 7233817     210     2020-07-11 28     2020  057900… 01         13
#>  4 Datos_2020… 7233837     210     2020-02-23 08     2020  056600… 01         15
#>  5 Datos_2020… 7233481     210     2020-10-20 41     2020  052660… 01         31
#>  6 Datos_2020… 7233482     210     2020-03-13 11     2020  050450… 35         15
#>  7 Datos_2020… 7233524     210     2020-11-18 46     2020  055790… 02         36
#>  8 Datos_2020… 7233531     210     2020-08-17 33     2020  056150… 01         27
#>  9 Datos_2020… 7233591     210     2020-05-09 18     2020  053900… 01          5
#> 10 Datos_2020… 7233682     210     2020-12-20 50     2020  056150… 01          5
#> # ℹ 1,700 more rows
#> # ℹ 65 more variables: uni_med <dbl>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …
geo_filtro(data_event = data_limpia, dpto = "05", mpio = "05001")
#> # A tibble: 504 × 74
#>    particion   consecutive cod_eve fec_not    semana ano   cod_pre cod_sub  edad
#>    <chr>       <chr>       <chr>   <date>     <chr>  <chr> <chr>   <chr>   <dbl>
#>  1 Datos_2020… 7233481     210     2020-10-20 41     2020  052660… 01         31
#>  2 Datos_2020… 7233755     210     2020-03-03 10     2020  050010… 08         15
#>  3 Datos_2020… 7345121     210     2020-09-10 36     2020  050010… 10         50
#>  4 Datos_2020… 7345145     210     2020-04-04 14     2020  050011… 07         33
#>  5 Datos_2020… 7234649     210     2020-02-06 05     2020  050010… 01         57
#>  6 Datos_2020… 7317731     210     2020-10-03 40     2020  050010… 01         23
#>  7 Datos_2020… 7317853     210     2020-01-22 03     2020  110013… 02          9
#>  8 Datos_2020… 7249163     210     2020-09-23 38     2020  052660… 03         66
#>  9 Datos_2020… 7249172     210     2020-01-18 02     2020  050010… 17         66
#> 10 Datos_2020… 7314789     210     2020-01-08 02     2020  050011… 01         27
#> # ℹ 494 more rows
#> # ℹ 65 more variables: uni_med <dbl>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …
geo_filtro(data_event = data_limpia, dpto = 05, mpio = 05001)
#> # A tibble: 504 × 74
#>    particion   consecutive cod_eve fec_not    semana ano   cod_pre cod_sub  edad
#>    <chr>       <chr>       <chr>   <date>     <chr>  <chr> <chr>   <chr>   <dbl>
#>  1 Datos_2020… 7233481     210     2020-10-20 41     2020  052660… 01         31
#>  2 Datos_2020… 7233755     210     2020-03-03 10     2020  050010… 08         15
#>  3 Datos_2020… 7345121     210     2020-09-10 36     2020  050010… 10         50
#>  4 Datos_2020… 7345145     210     2020-04-04 14     2020  050011… 07         33
#>  5 Datos_2020… 7234649     210     2020-02-06 05     2020  050010… 01         57
#>  6 Datos_2020… 7317731     210     2020-10-03 40     2020  050010… 01         23
#>  7 Datos_2020… 7317853     210     2020-01-22 03     2020  110013… 02          9
#>  8 Datos_2020… 7249163     210     2020-09-23 38     2020  052660… 03         66
#>  9 Datos_2020… 7249172     210     2020-01-18 02     2020  050010… 17         66
#> 10 Datos_2020… 7314789     210     2020-01-08 02     2020  050011… 01         27
#> # ℹ 494 more rows
#> # ℹ 65 more variables: uni_med <dbl>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …
geo_filtro(data_event = data_limpia, dpto = 05, mpio = 001)
#> # A tibble: 504 × 74
#>    particion   consecutive cod_eve fec_not    semana ano   cod_pre cod_sub  edad
#>    <chr>       <chr>       <chr>   <date>     <chr>  <chr> <chr>   <chr>   <dbl>
#>  1 Datos_2020… 7233481     210     2020-10-20 41     2020  052660… 01         31
#>  2 Datos_2020… 7233755     210     2020-03-03 10     2020  050010… 08         15
#>  3 Datos_2020… 7345121     210     2020-09-10 36     2020  050010… 10         50
#>  4 Datos_2020… 7345145     210     2020-04-04 14     2020  050011… 07         33
#>  5 Datos_2020… 7234649     210     2020-02-06 05     2020  050010… 01         57
#>  6 Datos_2020… 7317731     210     2020-10-03 40     2020  050010… 01         23
#>  7 Datos_2020… 7317853     210     2020-01-22 03     2020  110013… 02          9
#>  8 Datos_2020… 7249163     210     2020-09-23 38     2020  052660… 03         66
#>  9 Datos_2020… 7249172     210     2020-01-18 02     2020  050010… 17         66
#> 10 Datos_2020… 7314789     210     2020-01-08 02     2020  050011… 01         27
#> # ℹ 494 more rows
#> # ℹ 65 more variables: uni_med <dbl>, nacionalidad <chr>,
#> #   nombre_nacionalidad <chr>, sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>,
#> #   cod_mun_o <chr>, area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>,
#> #   per_etn <chr>, gru_pob <lgl>, nom_grupo <chr>, estrato <chr>,
#> #   gp_discapa <chr>, gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>,
#> #   gp_gestan <chr>, sem_ges <chr>, gp_indigen <chr>, gp_pobicfb <chr>, …
geo_filtro(data_event = data_limpia, dpto = "bogota dc", mpio = "bogota dc")
#> # A tibble: 0 × 74
#> # ℹ 74 variables: particion <chr>, consecutive <chr>, cod_eve <chr>,
#> #   fec_not <date>, semana <chr>, ano <chr>, cod_pre <chr>, cod_sub <chr>,
#> #   edad <dbl>, uni_med <dbl>, nacionalidad <chr>, nombre_nacionalidad <chr>,
#> #   sexo <chr>, cod_pais_o <chr>, cod_dpto_o <chr>, cod_mun_o <chr>,
#> #   area <chr>, ocupacion <chr>, tip_ss <chr>, cod_ase <chr>, per_etn <chr>,
#> #   gru_pob <lgl>, nom_grupo <chr>, estrato <chr>, gp_discapa <chr>,
#> #   gp_desplaz <chr>, gp_migrant <chr>, gp_carcela <chr>, gp_gestan <chr>, …