Limpiar fechas de los datos de una enfermedad o evento

Función que limpia y estandariza las fechas de los datos de una enfermedad o evento

Uso

limpiar_fecha_event(
  data_event,
  year,
  format_fecha = "%Y-%m-%d",
  col_fecha = "ini_sin",
  col_comp = NULL
)

Argumentos

data_event

Un `data.frame` que contiene los datos de una enfermedad o evento

year

Un `numeric` (numerico) o `character` (cadena de caracteres) que contiene el año de los datos de una enfermedad o evento

format_fecha

Un `character` (cadena de caracteres) que contiene el formato deseado de fecha; su valor por defecto es "%AAAA-%MM-%DD"

col_fecha

Un `character` (cadena de caracteres) que contiene el nombre de la columna con la fecha que se desea limpiar en los datos de la enfermedad o evento

col_comp

Un `character` (cadena de caracteres) que contiene el nombre de la columna con la cual se va a comparar la columna `nomb_col` para limpiarla, estandarizarla o aplicar las reglas definidas

Valor

Un `data.frame` con las fechas limpias

Ejemplos

data(dengue2020)
data_limpia <- limpiar_data_sivigila(data_event = dengue2020)
limpiar_fecha_event(data_event = data_limpia,
                    year = 2020,
                    format_fecha = "%Y-%m-%d",
                    col_fecha = "ini_sin",
                    col_comp = "fec_hos")
#> # A tibble: 0 × 73
#> # ℹ 73 variables: consecutive <chr>, cod_eve <chr>, fec_not <date>,
#> #   semana <chr>, ano <chr>, cod_pre <chr>, cod_sub <chr>, edad <dbl>,
#> #   uni_med <dbl>, nacionalidad <chr>, nombre_nacionalidad <chr>, sexo <chr>,
#> #   cod_pais_o <chr>, cod_dpto_o <chr>, cod_mun_o <chr>, area <chr>,
#> #   ocupacion <chr>, tip_ss <chr>, cod_ase <chr>, per_etn <chr>, gru_pob <chr>,
#> #   nom_grupo <lgl>, estrato <chr>, gp_discapa <chr>, gp_desplaz <chr>,
#> #   gp_migrant <chr>, gp_carcela <chr>, gp_gestan <chr>, sem_ges <lgl>, …