Importar las enfermedades y años disponibles disposibles para su descarga desde los microdatos de SIVIGILA

Función que obtiene las enfermedades y los años disponibles para su descarga desde los microdatos de SIVIGILA

Usage

list_events()

Value

Una list con las enfermedades y los años disponibles para su descarga desde los microdatos de SIVIGILA

Examples

list_events()
#>                                                      enfermedad
#> 1                                             ACCIDENTE OFIDICO
#> 2  AGRESIONES POR ANIMALES POTENCIALMENTE TRANSMISORES DE RABIA
#> 3                                          ANOMALIAS CONGENITAS
#> 4                                            BAJO PESO AL NACER
#> 5                            CÁNCER DE LA MAMA Y CUELLO UTERINO
#> 6                                               CANCER INFANTIL
#> 7                                                        CHAGAS
#> 8                                                   CHIKUNGUNYA
#> 9                                                        DENGUE
#> 10                                                 DENGUE GRAVE
#> 11                                                     DIFTERIA
#> 12                   ENCEFALITIS DEL NILO OCCIDENTAL EN HUMANOS
#> 13                      ENCEFALITIS EQUINA DEL OESTE EN HUMANOS
#> 14                     ENCEFALITIS EQUINA VENEZOLANA EN HUMANOS
#> 15                                       ENDOMETRITIS PUERPERAL
#> 16                               ENFERMEDADES HUERFANAS - RARAS
#> 17                            ESI - IRAG (VIGILANCIA CENTINELA)
#> 18                       EVENTO ADVERSO SEGUIDO A LA VACUNACION
#> 19                                              FIEBRE AMARILLA
#> 20                               FIEBRE TIFOIDEA Y PARATIFOIDEA
#> 21                                                  HEPATITIS A
#> 22                                                  HEPATITIS C
#> 23                                     HIPOTIROIDISMO CONGENITO
#> 24                            INFECCION ASOCIADA A DISPOSITIVOS
#> 25            INFECCION RESPIRATORIA AGUDA GRAVE IRAG INUSITADA
#> 26                                    INTOXICACION POR FARMACOS
#> 27                             INTOXICACION POR METALES PESADOS
#> 28                                     INTOXICACION POR METANOL
#> 29           INTOXICACION POR MONOXIDO DE CARBONO Y OTROS GASES
#> 30                   INTOXICACION POR OTRAS SUSTANCIAS QUIMICAS
#> 31                                 INTOXICACION POR PLAGUICIDAS
#> 32                                   INTOXICACION POR SOLVENTES
#> 33                     INTOXICACION POR SUSTANCIAS PSICOACTIVAS
#> 34                                                          ISO
#> 35                                        LEISHMANIASIS CUTANEA
#> 36                                         LEISHMANIASIS MUCOSA
#> 37                                       LEISHMANIASIS VISCERAL
#> 38                                                        LEPRA
#> 39                                                LEPTOSPIROSIS
#> 40                           LEUCEMIA AGUDA PEDIATRICA LINFOIDE
#> 41                           LEUCEMIA AGUDA PEDIATRICA MIELOIDE
#> 42                             MALARIA ASOCIADA (FORMAS MIXTAS)
#> 43                                           MALARIA COMPLICADA
#> 44                                           MALARIA FALCIPARUM
#> 45                                                MALARIA VIVAX
#> 46                                            MENINGITIS  OTROS
#> 47                                     MENINGITIS MENINGOCÓCICA
#> 48                        MENINGITIS POR HAEMOPHILUS INFLUENZAE
#> 49                                     MENINGITIS POR NEUMOCOCO
#> 50                                       MENINGITIS TUBERCULOSA
#> 51                                   MORBILIDAD MATERNA EXTREMA
#> 52                                           MORBILIDAD POR IRA
#> 53                                           MORTALIDAD MATERNA
#> 54                       MORTALIDAD PERINATAL Y NEONATAL TARDIA
#> 55                                        MORTALIDAD POR DENGUE
#> 56                                  MORTALIDAD POR DESNUTRICION
#> 57                                  MORTALIDAD POR EDA 0-4 AÑOS
#> 58                                           MORTALIDAD POR IRA
#> 59                                       MORTALIDAD POR MALARIA
#> 60                 PARALISIS FLACIDA AGUDA (MENORES DE 15 AÑOS)
#> 61                                                  PAROTIDITIS
#> 62                                                      RUBEOLA
#> 63                                                    SARAMPION
#> 64                                SINDROME DE RUBEOLA CONGENITA
#> 65                                           TETANOS ACCIDENTAL
#> 66                                             TETANOS NEONATAL
#> 67                                                   TOS FERINA
#>                                                                                                aa
#> 1        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 2        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 3        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 4                                      2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 5  2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2019, 2020, 2021
#> 6        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 7        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 8        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 9        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 10       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 11             2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2021
#> 12                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 13                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 14                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 15                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 16                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 17       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 18       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 19                         2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018
#> 20       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 21       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 22                                                                         2018, 2019, 2020, 2021
#> 23       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 24                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 25       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 26       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 27       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 28       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 29       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 30       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 31       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 32       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 33       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 34                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 35       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 36       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 37       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 38       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 39       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 40       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 41       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 42       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 43       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 44       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 45       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 46                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 47 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2019, 2020, 2021
#> 48 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2019, 2020, 2021
#> 49 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2019, 2020, 2021
#> 50       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 51       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 52                                     2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 53       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 54             2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 55       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 56                                           2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 57 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2019, 2020, 2021
#> 58       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 59       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 60             2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020
#> 61       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 62                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 63             2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020
#> 64                               2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017
#> 65       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 66       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021
#> 67       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021