Importar las enfermedades y años disponibles disposibles para su descarga desde los microdatos del SIVIGILA

Función que obtiene las enfermedades y los años disponibles para su descarga desde los microdatos del SIVIGILA

Uso

list_events()

Valor

Una `list` con las enfermedades y los años disponibles para su descarga desde los microdatos del SIVIGILA

Ejemplos

# \donttest{
list_events()
#>                                                               enfermedad
#> 1                                                      Accidente Ofídico
#> 2           Agresiones Por Animales Potencialmente Transmisores De Rabia
#> 3                                                   Anomalías Congénitas
#> 4                                                     Bajo Peso Al Nacer
#> 7                                                                 Chagas
#> 8                                                             Chikunguya
#> 5                                     Cáncer De La Mama Y Cuello Uterino
#> 6                                                        Cáncer Infantil
#> 9                                                                 Dengue
#> 10                                                          Dengue Grave
#> 11                                     Esi - Irag (Vigilancia Centinela)
#> 12                        Evento Adverso Grave Posterior A La Vacunación
#> 13                                                    Exposición A Flúor
#> 14                                                       Fiebre Amarilla
#> 15                                        Fiebre Tifoidea Y Paratifoidea
#> 16                                                           Hepatitis A
#> 17                      Hepatitis B, C Y Coinfección Hepatitis B Y Delta
#> 18                                                           Hepatitis C
#> 19                                              Hipotiroidismo Congénito
#> 20                     Infección Respiratoria Aguda Grave Irag Inusitada
#> 67                                                   Intento De Suicidio
#> 21                                                Intoxicación Por Gases
#> 22                                         Intoxicación Por Medicamentos
#> 23                                      Intoxicación Por Metales Pesados
#> 24                                              Intoxicación Por Metanol
#> 25                            Intoxicación Por Otras Sustancias Químicas
#> 26                                          Intoxicación Por Plaguicidas
#> 27                                            Intoxicación Por Solventes
#> 28                              Intoxicación Por Sustancias Psicoactivas
#> 29                                                 Leishmaniasis Cutánea
#> 30                                                  Leishmaniasis Mucosa
#> 31                                                Leishmaniasis Visceral
#> 32                                                                 Lepra
#> 33                                                         Leptospirosis
#> 34                                     Lesiones Por Pólvora Y Explosivos
#> 35                                    Leucemia Aguda Pediátrica Linfoide
#> 36                                    Leucemia Aguda Pediátrica Mieloide
#> 66                                                               Malaria
#> 37                                      Malaria Asociada (Formas Mixtas)
#> 38                                                    Malaria Complicada
#> 39                                                    Malaria Falciparum
#> 40                                                         Malaria Vivax
#> 41                                              Meningitis Meningocócica
#> 42                                 Meningitis Por Haemophilus Influenzae
#> 43                                              Meningitis Por Neumococo
#> 44                                                Meningitis Tuberculosa
#> 45                                                    Morbilidad Por Eda
#> 46                                                    Morbilidad Por Ira
#> 47                                                    Mortalidad Materna
#> 48                                Mortalidad Perinatal Y Neonatal Tardía
#> 49                                                 Mortalidad Por Dengue
#> 50                                           Mortalidad Por Desnutrición
#> 51                                           Mortalidad Por Eda 0-4 Años
#> 52                                                    Mortalidad Por Ira
#> 53                                                Mortalidad Por Malaria
#> 54                                                           Parotiditis
#> 55                                                          Rabia Humana
#> 56                                                             Sarampión
#> 59                                                            Tos Ferina
#> 60                                                               Tracoma
#> 61                                           Tuberculosis Extra Pulmonar
#> 62                                       Tuberculosis Farmacorresistente
#> 63                                                 Tuberculosis Pulmonar
#> 57                                                    Tétanos Accidental
#> 58                                                      Tétanos Neonatal
#> 64                                                   Varicela Individual
#> 65 Vigilancia En Salud Pública De La Violencia De Género E Intrafamiliar
#> 68                                                                  Zika
#>                                                                                                      aa
#> 1        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 2        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 3                          2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 4                                      2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 7                                      2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 8                                                  2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 5                                                                    2016, 2017, 2018, 2019, 2020, 2021
#> 6                                                  2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 9        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 10       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 11             2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 12       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 13                                                       2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019
#> 14                                                                   2007, 2008, 2009, 2013, 2016, 2018
#> 15       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 16       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 17       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 18                                                 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 19       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 20                   2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 67                                                                                                     
#> 21                         2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 22       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 23       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 24       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 25                                     2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 26       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 27       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 28                         2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 29       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 30       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 31       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 32       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 33       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 34                                                       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014
#> 35             2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 36             2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 66                                                                                                     
#> 37       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 38       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 39       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 40       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 41       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 42       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 43       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 44       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 45                                                 2013, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 46                                     2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 47       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 48       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 49       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 50                                           2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 51       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 52       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 53 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 54       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 55                                                 2007, 2008, 2009, 2010, 2012, 2015, 2016, 2017, 2020
#> 56                                                             2011, 2012, 2013, 2015, 2018, 2019, 2020
#> 59       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 60                                                                               2017, 2018, 2019, 2022
#> 61       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 62                               2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 63       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 57       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 58                   2007, 2008, 2009, 2010, 2012, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 64       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 65                                     2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 68                                                                                                     
# }