Importar las enfermedades y años disponibles disposibles para su descarga desde los microdatos del SIVIGILA

Función que obtiene las enfermedades y los años disponibles para su descarga desde los microdatos del SIVIGILA

Uso

list_events()

Valor

Una `list` con las enfermedades y los años disponibles para su descarga desde los microdatos del SIVIGILA

Ejemplos

# \donttest{
list_events()
#>                                                               enfermedad
#> 1                                                      Accidente Ofídico
#> 2           Agresiones Por Animales Potencialmente Transmisores De Rabia
#> 3                                                   Anomalías Congénitas
#> 4                                                     Bajo Peso Al Nacer
#> 7                                                                 Chagas
#> 8                                                             Chikunguya
#> 5                                     Cáncer De La Mama Y Cuello Uterino
#> 6                                                        Cáncer Infantil
#> 9                                                                 Dengue
#> 10                                                          Dengue Grave
#> 11                                     Esi - Irag (Vigilancia Centinela)
#> 12                                                         Eta Colectivo
#> 13                        Evento Adverso Grave Posterior A La Vacunación
#> 14                                                    Exposición A Flúor
#> 15                                                       Fiebre Amarilla
#> 16                                        Fiebre Tifoidea Y Paratifoidea
#> 17                                                           Hepatitis A
#> 18                      Hepatitis B, C Y Coinfección Hepatitis B Y Delta
#> 19                                                           Hepatitis C
#> 20                                              Hipotiroidismo Congénito
#> 21                     Infección Respiratoria Aguda Grave Irag Inusitada
#> 69                                                   Intento De Suicidio
#> 22                                                Intoxicación Por Gases
#> 23                                         Intoxicación Por Medicamentos
#> 24                                      Intoxicación Por Metales Pesados
#> 25                                              Intoxicación Por Metanol
#> 26                            Intoxicación Por Otras Sustancias Químicas
#> 27                                          Intoxicación Por Plaguicidas
#> 28                                            Intoxicación Por Solventes
#> 29                              Intoxicación Por Sustancias Psicoactivas
#> 30                                                 Leishmaniasis Cutánea
#> 31                                                  Leishmaniasis Mucosa
#> 32                                                Leishmaniasis Visceral
#> 33                                                                 Lepra
#> 34     Lesiones Por Artefactos Explosivos (Pólvora Y Minas Antipersonal)
#> 35                                     Lesiones Por Pólvora Y Explosivos
#> 36                                    Leucemia Aguda Pediátrica Linfoide
#> 37                                    Leucemia Aguda Pediátrica Mieloide
#> 68                                                               Malaria
#> 38                                      Malaria Asociada (Formas Mixtas)
#> 39                                                    Malaria Complicada
#> 40                                                    Malaria Falciparum
#> 41                                                       Malaria Malarie
#> 42                                                         Malaria Vivax
#> 43                                              Meningitis Meningocócica
#> 44                                 Meningitis Por Haemophilus Influenzae
#> 45                                              Meningitis Por Neumococo
#> 46                                                Meningitis Tuberculosa
#> 47                                                    Morbilidad Por Eda
#> 48                                                    Morbilidad Por Ira
#> 49                                                    Mortalidad Materna
#> 50                                Mortalidad Perinatal Y Neonatal Tardía
#> 51                                                 Mortalidad Por Dengue
#> 52                                           Mortalidad Por Desnutrición
#> 53                                           Mortalidad Por Eda 0-4 Años
#> 54                                                    Mortalidad Por Ira
#> 55                                                Mortalidad Por Malaria
#> 56                                                           Parotiditis
#> 57                                                               Rubeola
#> 58                                                             Sarampión
#> 61                                                            Tos Ferina
#> 62                                                               Tracoma
#> 63                                           Tuberculosis Extra Pulmonar
#> 64                                       Tuberculosis Farmacorresistente
#> 65                                                 Tuberculosis Pulmonar
#> 59                                                    Tétanos Accidental
#> 60                                                      Tétanos Neonatal
#> 66                                                   Varicela Individual
#> 67 Vigilancia En Salud Pública De La Violencia De Género E Intrafamiliar
#> 70                                                                  Zika
#>                                                                                                      aa
#> 1        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 2        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 3                          2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 4                                      2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 7                                      2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 8                                                  2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 5                                                                    2016, 2017, 2018, 2019, 2020, 2021
#> 6                                                  2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 9        2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 10       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 11             2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 12                         2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 13       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 14                                                       2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019
#> 15                                                                   2007, 2008, 2009, 2013, 2016, 2018
#> 16       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 17       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 18       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 19                                                 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 20       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 21                   2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 69                                                                                                     
#> 22                         2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 23       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 24       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 25       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 26                                     2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 27       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 28       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 29                         2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 30       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 31       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 32       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 33       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 34                                                 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 35                                                       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014
#> 36             2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 37             2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 68                                                                                                     
#> 38       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 39       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 40       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 41                                                       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014
#> 42       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 43       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 44       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 45       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 46       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 47                   2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 48                                     2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 49       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 50       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 51       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 52                                           2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 53       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 54       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 55 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 56       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 57                                                                         2007, 2008, 2009, 2011, 2012
#> 58                                                             2011, 2012, 2013, 2015, 2018, 2019, 2020
#> 61       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 62                                                                               2017, 2018, 2019, 2022
#> 63       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 64                               2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 65       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 59       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 60                   2007, 2008, 2009, 2010, 2012, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 66       2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 67                                     2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
#> 70                                                                                                     
# }