Obtener información geográfica de los datos de la enfermedad o evento

Función que obtiene la información geográfica de los datos de la enfermedad o evento

Usage

obtener_info_depts(dpto = NULL, munpio = NULL)

Arguments

dpto

Un character (cadena de caracteres) que contiene el nombre del departamento; su valor por defecto es NULL

munpio

Un character (cadena de caracteres) que contiene los datos del municipio; su valor por defecto es NULL

Value

Un data frame con la información geográfica de los datos de la enfermedad o evento

Examples

obtener_info_depts(dpto = "ANTIOQUIA")
#>     codigo_departamento codigo_municipio nombre_departamento
#> 1                    05             5001           ANTIOQUIA
#> 2                    05             5002           ANTIOQUIA
#> 3                    05             5004           ANTIOQUIA
#> 4                    05             5021           ANTIOQUIA
#> 5                    05             5030           ANTIOQUIA
#> 6                    05             5031           ANTIOQUIA
#> 7                    05             5034           ANTIOQUIA
#> 8                    05             5036           ANTIOQUIA
#> 9                    05             5038           ANTIOQUIA
#> 10                   05             5040           ANTIOQUIA
#> 11                   05             5042           ANTIOQUIA
#> 12                   05             5044           ANTIOQUIA
#> 13                   05             5045           ANTIOQUIA
#> 14                   05             5051           ANTIOQUIA
#> 15                   05             5055           ANTIOQUIA
#> 16                   05             5059           ANTIOQUIA
#> 17                   05             5079           ANTIOQUIA
#> 18                   05             5086           ANTIOQUIA
#> 19                   05             5088           ANTIOQUIA
#> 20                   05             5091           ANTIOQUIA
#> 21                   05             5093           ANTIOQUIA
#> 22                   05             5101           ANTIOQUIA
#> 23                   05             5107           ANTIOQUIA
#> 24                   05             5113           ANTIOQUIA
#> 25                   05             5120           ANTIOQUIA
#> 26                   05             5125           ANTIOQUIA
#> 27                   05             5129           ANTIOQUIA
#> 28                   05             5134           ANTIOQUIA
#> 29                   05             5138           ANTIOQUIA
#> 30                   05             5142           ANTIOQUIA
#> 31                   05             5145           ANTIOQUIA
#> 32                   05             5147           ANTIOQUIA
#> 33                   05             5148           ANTIOQUIA
#> 34                   05             5150           ANTIOQUIA
#> 35                   05             5154           ANTIOQUIA
#> 36                   05             5172           ANTIOQUIA
#> 37                   05             5190           ANTIOQUIA
#> 38                   05             5197           ANTIOQUIA
#> 39                   05             5206           ANTIOQUIA
#> 40                   05             5209           ANTIOQUIA
#> 41                   05             5212           ANTIOQUIA
#> 42                   05             5234           ANTIOQUIA
#> 43                   05             5237           ANTIOQUIA
#> 44                   05             5240           ANTIOQUIA
#> 45                   05             5250           ANTIOQUIA
#> 46                   05             5264           ANTIOQUIA
#> 47                   05             5266           ANTIOQUIA
#> 48                   05             5282           ANTIOQUIA
#> 49                   05             5284           ANTIOQUIA
#> 50                   05             5306           ANTIOQUIA
#> 51                   05             5308           ANTIOQUIA
#> 52                   05             5310           ANTIOQUIA
#> 53                   05             5313           ANTIOQUIA
#> 54                   05             5315           ANTIOQUIA
#> 55                   05             5318           ANTIOQUIA
#> 56                   05             5321           ANTIOQUIA
#> 57                   05             5347           ANTIOQUIA
#> 58                   05             5353           ANTIOQUIA
#> 59                   05             5360           ANTIOQUIA
#> 60                   05             5361           ANTIOQUIA
#> 61                   05             5364           ANTIOQUIA
#> 62                   05             5368           ANTIOQUIA
#> 63                   05             5376           ANTIOQUIA
#> 64                   05             5380           ANTIOQUIA
#> 65                   05             5390           ANTIOQUIA
#> 66                   05             5400           ANTIOQUIA
#> 67                   05             5411           ANTIOQUIA
#> 68                   05             5425           ANTIOQUIA
#> 69                   05             5440           ANTIOQUIA
#> 70                   05             5467           ANTIOQUIA
#> 71                   05             5475           ANTIOQUIA
#> 72                   05             5480           ANTIOQUIA
#> 73                   05             5483           ANTIOQUIA
#> 74                   05             5490           ANTIOQUIA
#> 75                   05             5495           ANTIOQUIA
#> 76                   05             5501           ANTIOQUIA
#> 77                   05             5541           ANTIOQUIA
#> 78                   05             5543           ANTIOQUIA
#> 79                   05             5576           ANTIOQUIA
#> 80                   05             5579           ANTIOQUIA
#> 81                   05             5585           ANTIOQUIA
#> 82                   05             5591           ANTIOQUIA
#> 83                   05             5604           ANTIOQUIA
#> 84                   05             5607           ANTIOQUIA
#> 85                   05             5615           ANTIOQUIA
#> 86                   05             5628           ANTIOQUIA
#> 87                   05             5631           ANTIOQUIA
#> 88                   05             5642           ANTIOQUIA
#> 89                   05             5647           ANTIOQUIA
#> 90                   05             5649           ANTIOQUIA
#> 91                   05             5652           ANTIOQUIA
#> 92                   05             5656           ANTIOQUIA
#> 93                   05             5658           ANTIOQUIA
#> 94                   05             5659           ANTIOQUIA
#> 95                   05             5660           ANTIOQUIA
#> 96                   05             5664           ANTIOQUIA
#> 97                   05             5665           ANTIOQUIA
#> 98                   05             5667           ANTIOQUIA
#> 99                   05             5670           ANTIOQUIA
#> 100                  05             5674           ANTIOQUIA
#> 101                  05             5679           ANTIOQUIA
#> 102                  05             5686           ANTIOQUIA
#> 103                  05             5690           ANTIOQUIA
#> 104                  05             5697           ANTIOQUIA
#> 105                  05             5736           ANTIOQUIA
#> 106                  05             5756           ANTIOQUIA
#> 107                  05             5761           ANTIOQUIA
#> 108                  05             5789           ANTIOQUIA
#> 109                  05             5790           ANTIOQUIA
#> 110                  05             5792           ANTIOQUIA
#> 111                  05             5809           ANTIOQUIA
#> 112                  05             5819           ANTIOQUIA
#> 113                  05             5837           ANTIOQUIA
#> 114                  05             5842           ANTIOQUIA
#> 115                  05             5847           ANTIOQUIA
#> 116                  05             5854           ANTIOQUIA
#> 117                  05             5856           ANTIOQUIA
#> 118                  05             5858           ANTIOQUIA
#> 119                  05             5861           ANTIOQUIA
#> 120                  05             5873           ANTIOQUIA
#> 121                  05             5885           ANTIOQUIA
#> 122                  05             5887           ANTIOQUIA
#> 123                  05             5890           ANTIOQUIA
#> 124                  05             5893           ANTIOQUIA
#> 125                  05             5895           ANTIOQUIA
#>              nombre_municipio tipo_municipio_isla_area_no_municipalizada
#> 1                    MEDELLÍN                                  Municipio
#> 2                   ABEJORRAL                                  Municipio
#> 3                    ABRIAQUÍ                                  Municipio
#> 4                  ALEJANDRÍA                                  Municipio
#> 5                       AMAGÁ                                  Municipio
#> 6                      AMALFI                                  Municipio
#> 7                       ANDES                                  Municipio
#> 8                 ANGELÓPOLIS                                  Municipio
#> 9                   ANGOSTURA                                  Municipio
#> 10                      ANORÍ                                  Municipio
#> 11      SANTA FÉ DE ANTIOQUIA                                  Municipio
#> 12                       ANZÁ                                  Municipio
#> 13                   APARTADÓ                                  Municipio
#> 14                  ARBOLETES                                  Municipio
#> 15                    ARGELIA                                  Municipio
#> 16                    ARMENIA                                  Municipio
#> 17                    BARBOSA                                  Municipio
#> 18                    BELMIRA                                  Municipio
#> 19                      BELLO                                  Municipio
#> 20                    BETANIA                                  Municipio
#> 21                    BETULIA                                  Municipio
#> 22             CIUDAD BOLÍVAR                                  Municipio
#> 23                    BRICEÑO                                  Municipio
#> 24                   BURITICÁ                                  Municipio
#> 25                    CÁCERES                                  Municipio
#> 26                    CAICEDO                                  Municipio
#> 27                     CALDAS                                  Municipio
#> 28                 CAMPAMENTO                                  Municipio
#> 29                CAÑASGORDAS                                  Municipio
#> 30                   CARACOLÍ                                  Municipio
#> 31                  CARAMANTA                                  Municipio
#> 32                     CAREPA                                  Municipio
#> 33       EL CARMEN DE VIBORAL                                  Municipio
#> 34                   CAROLINA                                  Municipio
#> 35                   CAUCASIA                                  Municipio
#> 36                  CHIGORODÓ                                  Municipio
#> 37                   CISNEROS                                  Municipio
#> 38                    COCORNÁ                                  Municipio
#> 39                 CONCEPCIÓN                                  Municipio
#> 40                  CONCORDIA                                  Municipio
#> 41                 COPACABANA                                  Municipio
#> 42                    DABEIBA                                  Municipio
#> 43                  DONMATÍAS                                  Municipio
#> 44                    EBÉJICO                                  Municipio
#> 45                   EL BAGRE                                  Municipio
#> 46                 ENTRERRÍOS                                  Municipio
#> 47                   ENVIGADO                                  Municipio
#> 48                   FREDONIA                                  Municipio
#> 49                   FRONTINO                                  Municipio
#> 50                    GIRALDO                                  Municipio
#> 51                  GIRARDOTA                                  Municipio
#> 52                GÓMEZ PLATA                                  Municipio
#> 53                    GRANADA                                  Municipio
#> 54                  GUADALUPE                                  Municipio
#> 55                     GUARNE                                  Municipio
#> 56                    GUATAPÉ                                  Municipio
#> 57                  HELICONIA                                  Municipio
#> 58                   HISPANIA                                  Municipio
#> 59                     ITAGÜÍ                                  Municipio
#> 60                    ITUANGO                                  Municipio
#> 61                     JARDÍN                                  Municipio
#> 62                     JERICÓ                                  Municipio
#> 63                    LA CEJA                                  Municipio
#> 64                LA ESTRELLA                                  Municipio
#> 65                 LA PINTADA                                  Municipio
#> 66                   LA UNIÓN                                  Municipio
#> 67                   LIBORINA                                  Municipio
#> 68                      MACEO                                  Municipio
#> 69                  MARINILLA                                  Municipio
#> 70                 MONTEBELLO                                  Municipio
#> 71                    MURINDÓ                                  Municipio
#> 72                     MUTATÁ                                  Municipio
#> 73                     NARIÑO                                  Municipio
#> 74                    NECOCLÍ                                  Municipio
#> 75                      NECHÍ                                  Municipio
#> 76                      OLAYA                                  Municipio
#> 77                      PEÑOL                                  Municipio
#> 78                      PEQUE                                  Municipio
#> 79                PUEBLORRICO                                  Municipio
#> 80              PUERTO BERRÍO                                  Municipio
#> 81                PUERTO NARE                                  Municipio
#> 82             PUERTO TRIUNFO                                  Municipio
#> 83                   REMEDIOS                                  Municipio
#> 84                     RETIRO                                  Municipio
#> 85                   RIONEGRO                                  Municipio
#> 86                SABANALARGA                                  Municipio
#> 87                   SABANETA                                  Municipio
#> 88                     SALGAR                                  Municipio
#> 89     SAN ANDRÉS DE CUERQUÍA                                  Municipio
#> 90                 SAN CARLOS                                  Municipio
#> 91              SAN FRANCISCO                                  Municipio
#> 92               SAN JERÓNIMO                                  Municipio
#> 93     SAN JOSÉ DE LA MONTAÑA                                  Municipio
#> 94          SAN JUAN DE URABÁ                                  Municipio
#> 95                   SAN LUIS                                  Municipio
#> 96  SAN PEDRO DE LOS MILAGROS                                  Municipio
#> 97         SAN PEDRO DE URABÁ                                  Municipio
#> 98                 SAN RAFAEL                                  Municipio
#> 99                  SAN ROQUE                                  Municipio
#> 100        SAN VICENTE FERRER                                  Municipio
#> 101             SANTA BÁRBARA                                  Municipio
#> 102        SANTA ROSA DE OSOS                                  Municipio
#> 103             SANTO DOMINGO                                  Municipio
#> 104              EL SANTUARIO                                  Municipio
#> 105                   SEGOVIA                                  Municipio
#> 106                    SONSÓN                                  Municipio
#> 107                  SOPETRÁN                                  Municipio
#> 108                   TÁMESIS                                  Municipio
#> 109                    TARAZÁ                                  Municipio
#> 110                     TARSO                                  Municipio
#> 111                  TITIRIBÍ                                  Municipio
#> 112                    TOLEDO                                  Municipio
#> 113                     TURBO                                  Municipio
#> 114                   URAMITA                                  Municipio
#> 115                     URRAO                                  Municipio
#> 116                  VALDIVIA                                  Municipio
#> 117                VALPARAÍSO                                  Municipio
#> 118                   VEGACHÍ                                  Municipio
#> 119                   VENECIA                                  Municipio
#> 120          VIGÍA DEL FUERTE                                  Municipio
#> 121                      YALÍ                                  Municipio
#> 122                   YARUMAL                                  Municipio
#> 123                   YOLOMBÓ                                  Municipio
#> 124                     YONDÓ                                  Municipio
#> 125                  ZARAGOZA                                  Municipio